% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_0d.R
\name{assess_0d}
\alias{assess_0d}
\title{Assess a model on a single data set with 0-dim}
\usage{
assess_0d(
  expr_mat,
  pheno_tbl,
  data_spec,
  model_spec,
  ass_spec_0d,
  quiet = FALSE,
  msg_prefix = ""
)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix. Rows are samples, columns
are genes, their identifiers are given as row and column names, respectively.}

\item{pheno_tbl}{tibble. The phenotype table. Observations are samples, columns
are variables.}

\item{data_spec}{DataSpec object specifying \code{expr_mat} and \code{pheno_tbl}. See the
constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec object. The model to be assessed. See the constructor
\code{ModelSpec()} for details.}

\item{ass_spec_0d}{AssSpec0d S3 object. The specifications for the 0-dim metric.
See the constructor \code{\link[=AssSpec0d]{AssSpec0d()}} for details.}

\item{quiet}{logical. Whether to suppress messages. Default is \code{FALSE}.}

\item{msg_prefix}{string. Prefix for messages. Default is \code{""}.}
}
\value{
numeric vector. For every split index the desired metric.
}
\description{
Map every model to a 0-dim metric (a single real number) and store
these metrics together with some more analytics across the splits in a sorted
tibble. The ground truth is time to event less or equal a cutoff.
}
\details{
The AssSpec0d S3 class is tailored for this function.
}
