% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_0d_center.R
\name{assess_0d_center}
\alias{assess_0d_center}
\title{Assess multiple models on a single data set with one single value}
\usage{
assess_0d_center(
  ass_spec_0d,
  data_spec,
  model_spec_list,
  cohorts = c("test", "train"),
  model_tree_mirror = c("models", "results"),
  quiet = FALSE
)
}
\arguments{
\item{ass_spec_0d}{AssSpec0d S3 object. See the constructor \code{\link[=AssSpec0d]{AssSpec0d()}} for more
details.}

\item{data_spec}{DataSpec object. Assess on this data set.}

\item{model_spec_list}{list of ModelSpec objects. Assess these models.}

\item{cohorts}{character vector, a subset of \code{c("train", "test")}. Assess on these
cohorts. Default is \code{c("train", "test")}.}

\item{model_tree_mirror}{character vector of length 2. If you want to store the
resulting tibbles, get the test-cohort tibble's file name by mirroring
\code{ass_spec_0d$file} according to \code{model_tree_mirror} (see \code{\link[=mirror_directory]{mirror_directory()}}).}

\item{quiet}{logical. Whether to suppress messages. Default is \code{FALSE}.}
}
\value{
A tibble. Every row holds the metric (and more analysis across the splits
like standard deviation, minimum, maximum value) for one model in \code{model_spec_list}
and every time cutoff specified for this model.
}
\description{
Map every model to a 0-dim metric (a single real number) and store
these metrics together with some more analytics across the splits in a sorted
tibble. At the core, we call \verb{[assess_0d()]}.
}
