% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_2d_center.R
\name{assess_2d_center}
\alias{assess_2d_center}
\title{Assess multiple models on a single data set}
\usage{
assess_2d_center(
  ass_2d_spec,
  model_spec_list,
  data_spec,
  cohorts = c("train", "test"),
  model_tree_mirror = c("models", "results"),
  comparison_plot = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{ass_2d_spec}{Ass2dSpec object. Specify the final comparison plot.
We derive the \code{Ass2dSpec} for the single plots in a reasonable way from it.}

\item{model_spec_list}{list of ModelSpec objects. Assess these models.}

\item{data_spec}{DataSpec object. Assess on this data set.}

\item{cohorts}{character vector, a subset of \code{c("train", "test")}. Assess on these
cohorts. Default is \code{c("train", "test")}.}

\item{model_tree_mirror}{character vector of lengtgh 2. This answers the question
where to store the generated assessment files.
\itemize{
\item For the train cohort, it is the model directory (where the fit lies as an rds
file).
\item For the test cohort, it is the same directory as for the train cohort, with one
exception: we mirror the file path according to \code{model_tree_mirror}, i.e., we
replace the first element of \code{model_tree_mirror} by the second element. Default
is \code{c("models", "results")}.
}}

\item{comparison_plot}{logical. Whether to generate the above mentioned comparison
plot. Default is \code{TRUE}.}

\item{quiet}{logical. Whether to suppress messages. Default is \code{FALSE}.}
}
\description{
Assess the performance of multiple models on a single data set.
For every
\itemize{
\item cohort (train and test),
\item model and
\item time cutoff this model specifies,
call \code{\link[=assess_2d]{assess_2d()}}. In addition, plot \code{x_metric} vs. \code{y_metric} attribute
of \code{Ass2dSpec} for all models in a sinlge plot ("comparison plot").
}
}
