% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_model.R
\name{assess_model}
\alias{assess_model}
\title{Assess a model on a data set in terms of classifying high-risk patients}
\usage{
assess_model(expr_mat, pheno_tbl, data_spec, model_spec, perf_plot_spec)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix. Rows are samples, columns
are genes, their identifiers are given as row and column names, respectively.}

\item{pheno_tbl}{tibble. The phenotype table. Observations are samples, columns
are variables.}

\item{data_spec}{DataSpec object specifying \code{expr_mat} and \code{pheno_tbl}. See the
constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec object. The model to be assessed. See the constructor
\code{ModelSpec()} for details.}

\item{perf_plot_spec}{PerfPlotSpec object. The specifications for the plot. See
the constructor \code{PerfPlotSpec()} for details. The \code{pfs_leq} attribute of
\code{perf_plot_spec} will override the same in \code{model_spec} if both are given.}
}
\value{
A named list of two tibbles:
\itemize{
\item \code{model_spec$name}: The performance measures for the model with three variables:
\code{perf_plot_spec$x_metric}, \code{perf_plot_spec$y_metric}, and \code{"cutoff"}.
\item \code{perf_plot_spec$benchmark}: The performance measures for the benchmark classifier
with the same three variables.
}
}
\description{
Given a data-model pair, assess the model's performance in terms of
classifying high-risk patients by plotting one binary-classifier characteristic
(e.g. prevalance) versus another one (e.g. precision) and comparing it to a
benchmark classifier (e.g. the International Prognostic Index (IPI) for DLBCL).
}
\details{
The assessment views every model as a binary classifier for high-risk (pfs <
\code{pfs_leq}) versus low-risk (pfs >= \code{pfs_leq}) patients (where \code{pfs_leq} is determined
as described above).
}
