% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_multiple_models.R
\name{assess_train_and_test}
\alias{assess_train_and_test}
\title{Call \code{\link[=assess_multiple_models]{assess_multiple_models()}} on training and test data}
\usage{
assess_train_and_test(
  model_spec_list,
  data_spec_train,
  data_spec_test,
  perf_plot_spec_train,
  model_tree_mirror = c("models", "results"),
  single_plots = TRUE,
  comparison_plot = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{model_spec_list}{list of ModelSpec objects. The models to be assessed.}

\item{data_spec_train}{DataSpec object for the training data.}

\item{data_spec_test}{DataSpec object for the test data.}

\item{perf_plot_spec_train}{PerfPlotSpec object for the training data. The same
PerfPlotSpec object will be used for the test data, except that the file path
will be mirrored to the results directory.}

\item{model_tree_mirror}{character vector of lengtgh 2. See \code{\link[=assess_multiple_models]{assess_multiple_models()}}.
Here, we also use it to mirror the \code{fname} attribute of \code{perf_plot_spec_train} to
obtain the \code{PerfPlotSpec} object for the test data.}

\item{single_plots}{logical. Whether to also generate individual plots for
every model-data pair. Default is \code{TRUE}.}

\item{comparison_plot}{logical. Whether to compare all models in a single plot.}

\item{quiet}{logical. Whether to suppress messages. Default is \code{FALSE}.}
}
\description{
Given data specifications for training and test data, a list
of model specifications and a plot specification for the training data, call
\code{assess_multiple_models()} twice: for the training data and, after mirroring
the comparison plot file path to the results directory, for the test data.
}
