% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass_scalar.R
\name{AssScalar}
\alias{AssScalar}
\title{An R6 class to assess a model with scalar metrics}
\description{
Assess how well a model can predict time to event less than a certain
threshold with a \emph{scalar} metric.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metrics}}{Assess the model for these scalar metrics. Check out the
initializer for possible choices.}

\item{\code{prev_range}}{For metrics that need thresholding only consider
thresholds that yield a prevalence in this range.}

\item{\code{confidence_level}}{Confidence level gamma, e.g. for confidence intervals.}

\item{\code{benchmark}}{Name and pivot time cutoff of the benchmark \code{Model}.}

\item{\code{round_digits}}{Round the results in tables to round_digits digits
after the point.}

\item{\code{file}}{Save the resulting tibble to this csv file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AssScalar-new}{\code{AssScalar$new()}}
\item \href{#method-AssScalar-assess}{\code{AssScalar$assess()}}
\item \href{#method-AssScalar-assess_center}{\code{AssScalar$assess_center()}}
\item \href{#method-AssScalar-clone}{\code{AssScalar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-new"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-new}{}}}
\subsection{Method \code{new()}}{
Construct an \code{AssScalar} R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$new(
  metrics = c("auc", "accuracy", "precision", "prevalence", "precision_ci_ll",
    "precision_ci_ul", "hr", "hr_ci_ll", "hr_ci_ul", "hr_p", "n_true", "perc_true",
    "n_samples", "logrank", "threshold"),
  prev_range = c(0, 1),
  confidence_level = 0.95,
  benchmark = NULL,
  file = NULL,
  round_digits = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metrics}}{character. Assess the model for these metrics. For
currently offered choices see "Usage". If you have a model with
non-binary output (like the linear predictor of a Cox model), we choose a threshold
by maximizing the left-most metric in \code{metrics} \emph{that is made for
classifiers with binary output} (e.g. precision within \code{prev_range} below).
If this cannot be done reasonably, throw an error. Make sure that \code{hr} precedes
\code{hr_ci_ll}, \code{hr_ci_ul} and \code{hr_p} in \code{metrics}; \code{precision_ci_ll} must precede
\code{precision_ci_ul}.}

\item{\code{prev_range}}{numeric numeric vector of length 2. For metrics that
need thresholding only consider thresholds that yield a prevalence in
this range.}

\item{\code{confidence_level}}{numeric. The confidence level gamma (e.g. for confidence
intervals).}

\item{\code{benchmark}}{list or \code{NULL}. If not NULL, it is a list with names
\itemize{
\item \code{"name"}: the \code{name} attribute of the benchmark \code{Model} in the \code{model_list} parameter
of the \code{assess()} and \code{assess_center()} method,
\item \code{"prev_range"}: An extra value for the \code{prev_range} attribute used for the
benchmark \code{Model}. Often, we need a higher prevalence for our, new models to gain
statistical power and be able to significantly outperform the benchmark.
}}

\item{\code{file}}{string or NULL. If not \code{NULL}, save the resulting tibble to this csv file.}

\item{\code{round_digits}}{numeric. The number of digits to round the results to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AssScalar} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-assess"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-assess}{}}}
\subsection{Method \code{assess()}}{
Assess a \emph{single} model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$assess(data, model, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data. Data must already be read in
and its \code{cohort} attribute set.}

\item{\code{model}}{Model object. Assess this model.}

\item{\code{quiet}}{logical. Whether to suppress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
named numeric vector. The calculated metrics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-assess_center"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-assess_center}{}}}
\subsection{Method \code{assess_center()}}{
Wrap \code{assess()} to assess \emph{multiple} models and store the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$assess_center(data, model_list, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data. The \code{cohort} attribute of \code{data} must be
set.}

\item{\code{model_list}}{list of Model objects. Assess these models.}

\item{\code{quiet}}{logical. Whether to suppress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble of shape \verb{(length(model_list) x length(metrics))}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
