% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass_spec_0d.R
\name{AssSpec0d}
\alias{AssSpec0d}
\title{An S3 class specifying how to assess with a 0-dim metric (like AUC)}
\usage{
AssSpec0d(
  metric,
  pivot_time_cutoff,
  lambda = "lambda.min",
  benchmark = NULL,
  file = NULL,
  round_digits = 3
)
}
\arguments{
\item{metric}{character. Name of the function used to calculate the metric. It must
take two numeric vectors of the same length as arguments:
\enumerate{
\item \code{predicted}, the predictions by a model, higher values indicate a more positive
prediction,
\item \code{actual}, the true values with the positive class coded as 1, the negative one as 0.
}}

\item{pivot_time_cutoff}{numeric. Time-to-event threshold that divides samples into a
high/low-risk (time to event below/above \code{pivot_time_cutoff}) group. Model performance
will be measured in terms of how well it can separate these two groups as a binary
classifier. Default is NULL, in which case the \code{pivot_time_cutoff} of an accompanying
DataSpec is used.}

\item{lambda}{numeric or string. The lambda regularization parameter of the model to
evaluate. See \code{\link[=prepare_and_predict]{prepare_and_predict()}} for details.}

\item{benchmark}{string. Column in the test pheno holding the numeric benchmark values.
Default is \code{"ipi"} (international prognostic index for DLBCL).}

\item{file}{string. The name of the csv file to save the results to. If \code{NULL},
the results are not saved. If you supply an AssSpec0d object to \code{assess_0d_center()},
specify the path for the \emph{train} cohort.}

\item{round_digits}{numeric. The number of digits to round the results to. Default is \code{3}.}
}
\value{
An AssSpec0d S3 object.
}
\description{
An S3 class specifying how to assess with a 0-dim metric (like AUC)
}
