% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass_2d_spec.R
\name{AssSpec2d}
\alias{AssSpec2d}
\title{Create a AssSpec2d object}
\usage{
AssSpec2d(
  file,
  x_metric,
  y_metric,
  pivot_time_cutoff = NULL,
  lambda = "lambda.min",
  benchmark = NULL,
  ci_level = 0.95,
  fellow_csv = FALSE,
  scores_plot = FALSE,
  show_plots = FALSE,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  xlim = c(-Inf, Inf),
  ylim = c(-Inf, Inf),
  smooth_method = NULL,
  smooth_benchmark = FALSE,
  smooth_se = FALSE,
  scale_x = "identity",
  scale_y = "identity",
  vline = NULL,
  hline = NULL,
  text = NULL,
  alpha = 0.5,
  width = 7,
  height = 4,
  colors = NULL,
  units = "in"
)
}
\arguments{
\item{file}{string. The name of the file to save the plot to.}

\item{x_metric}{string. The name of the performance measure to be plotted on the x-axis.
All measures that can be passed to the \code{x.measure} parameter of \code{\link[ROCR:performance]{ROCR::performance()}} are
valid.}

\item{y_metric}{string. The name of the performance measure to be plotted on the y-axis.
\itemize{
\item All measures that can be passed to the \code{measure} parameter of \code{\link[ROCR:performance]{ROCR::performance()}} are
valid.
\item \code{"logrank"} for the p-values of a logrank test. In this case, \code{x_metric} must be
\code{"prevalence"} or \code{"rpp"}.
}}

\item{pivot_time_cutoff}{numeric. Time-to-event threshold that divides samples into a
high/low-risk (time to event below/above \code{pivot_time_cutoff}) group. Model performance
will be measured in terms of how well it can separate these two groups as a binary
classifier. Default is NULL, in which case the \code{pivot_time_cutoff} of an accompanying
DataSpec is used.}

\item{lambda}{numeric or string. The lambda regularization parameter of the model to
evaluate. See \code{\link[=prepare_and_predict]{prepare_and_predict()}} for details. Default is \code{"lambda.min"}.}

\item{benchmark}{string. Column in the test pheno holding the numeric benchmark values.
Default is \code{"ipi"} (international prognostic index for DLBCL).}

\item{ci_level}{numeric in [0, 1]. The level used to calculate confidence intervals.
Default is \code{0.95}.}

\item{fellow_csv}{logical. If passed to \code{\link[=assess_2d_center]{assess_2d_center()}}, whether to also create a
csv file for every model-data pair. Default is \code{TRUE}.}

\item{scores_plot}{logical. Display the ordered scores output by the model in a scatter
plot. Default is \code{TRUE}.}

\item{show_plots}{logical. Whether to show the plots after creation in an interactive
graphics device. Default is \code{FALSE}.}

\item{title}{string. The title of the plot. Default is \code{NULL}.}

\item{x_lab, y_lab}{string. Axis labels. Default is \code{x_metric} and \code{y_metric}, respectively.}

\item{xlim, ylim}{numeric vector of length 2. The limits for both axes. Default is
\code{c(-Inf, Inf)}, i.e. no contraints.}

\item{smooth_method}{string or function. Smooth method to plot an additional smoothed graph.
If \code{NULL}, no smoothing. Else we pass \code{smooth_method} as the \code{method} parameter to
\code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. Default is \code{"loess"}.}

\item{smooth_benchmark}{logical. Whether to also smooth the benchmark data. Default is
\code{FALSE}.}

\item{smooth_se}{logical. Whether to add standard error bands to the smoothed lines.}

\item{scale_x, scale_y}{string or transformation object (see \code{\link[scales:new_transform]{scales::trans_new}} for the
latter). The scale of the axes, we will pass them to the \code{trans} paramter of
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}, [\verb{ggplot2::scale_y_continuous()], respectively.  Default is }"identity"`.}

\item{vline, hline}{list or NULL. Vertical/horizontal lines to be added to the plot. A list
holding the arguments to pass to \code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}} and \code{\link[ggplot2:geom_abline]{ggplot2::geom_hline()}},
respectively. Default is \code{NULL}.}

\item{text}{list or NULL. Text label added to the plot. A list holding the arguments to
pass to \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}. Default is \code{NULL}.}

\item{alpha}{numeric in [0, 1]. The alpha value for the points and lines in the
plot.}

\item{width}{numeric. The width of the plot in \code{units}. Default is \code{7}.}

\item{height}{numeric. The height of the plot in \code{units}. Default is \code{4}.}

\item{colors}{character vector. The colors to be used for the different models.
Default is \code{NULL}, which means that the default colors of \code{ggplot2} will be used.}

\item{units}{string. The units of \code{width} and \code{height}. Default is \code{"in"} (inches).}
}
\value{
An AssSpec2d S3 object.
}
\description{
A AssSpec2d object holds all the info on how to assess how well a
model filters high-risk patients. Its base object is a list. The core of the
assessment is a scatter plot of two performance measures. Moreover, a AssSpec2d
object holds the commands (usually bools) on whether do more assessments (that
require considerably less specifications).
}
