% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_tbl_columns_exist}
\alias{check_tbl_columns_exist}
\title{Check if a tibble has certain columns}
\usage{
check_tbl_columns_exist(tbl, tbl_name, col_names)
}
\arguments{
\item{tbl}{tibble. The tibble to check.}

\item{tbl_name}{string. How to refer to the tibble in an error message.}

\item{col_names}{vector of strings. The names of the columns to check for.}
}
\value{
NULL. Throws an error if one of the columns is not in the tibble.
}
\description{
Check if a tibble has certain columns
}
