% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Data}
\alias{Data}
\title{An R6 class for the data}
\description{
A Data object holds the phenotype and expression data belonging to
a data set. It specifies the names of important features in the columns, reads from
and writes to csv files and takes care of splitting the data into train and test
cohort.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{How to refer to the data in plots and messages.}

\item{\code{directory}}{Directory where the expression and pheno csv files lie.}

\item{\code{train_prop}}{Proportion of samples in the training cohort.}

\item{\code{expr_mat}}{The expression matrix.}

\item{\code{pheno_tbl}}{The phenotype tibble.}

\item{\code{expr_file}}{Name of the expression csv file inside \code{directory}.}

\item{\code{pheno_file}}{Name of the pheno data csv inside \code{directory}.}

\item{\code{cohort}}{The cohort of the data, train or test.}

\item{\code{patient_id_col}}{The name of the column in the pheno data that holds
the patient identifiers.}

\item{\code{time_to_event_col}}{The name of the column in the pheno data that holds the
time-to-event values.}

\item{\code{event_col}}{The name of the column in the pheno data that holds
the event status encoded as 1 = occurrence, 0 = censoring.}

\item{\code{benchmark_col}}{The name of the column in the pheno data that holds the
benchmark risk score (like the IPI).}

\item{\code{gene_id_col}}{The name of the column in the expression data that holds
the gene identifiers.}

\item{\code{split_col_prefix}}{Column-name prefix for those columns holding splits into
train and test cohort.}

\item{\code{pivot_time_cutoff}}{Pivotal time cutoff for the analysis and, explicitly,
for splitting the data into the train and test cohort.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Data-new}{\code{Data$new()}}
\item \href{#method-Data-read}{\code{Data$read()}}
\item \href{#method-Data-clone}{\code{Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-new"></a>}}
\if{latex}{\out{\hypertarget{method-Data-new}{}}}
\subsection{Method \code{new()}}{
A DataSpec object specifies the location and format of the expression
and pheno data of a single data set. This enables reading and preparing the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$new(
  name,
  directory,
  train_prop,
  pivot_time_cutoff = 2,
  expr_file = "expr.csv",
  pheno_file = "pheno.csv",
  cohort = NULL,
  patient_id_col = "patient_id",
  time_to_event_col = "pfs_years",
  split_col_prefix = "split_",
  event_col = "progression",
  benchmark_col = NULL,
  gene_id_col = "gene_id"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string. A telling name for the data set (e.g. \code{"Schmitz et al. 2018"}).}

\item{\code{directory}}{string. The directory where both expression and pheno csv
files lie.}

\item{\code{train_prop}}{numeric in (0, 1). The proportion of samples to draw for the
training cohort in every splitting run.}

\item{\code{pivot_time_cutoff}}{numeric or NULL. Pivotal time cutoff for the analysis and, explicitly,
for splitting the data into the train and test cohort: If a numeric in (0, 1), preserve
the the proportion of individuals below and above \code{time_cutoff} in both cohorts. Default
is \code{NULL}, i.e. no further constraints on splitting.}

\item{\code{expr_file}}{string. The name of the expression csv file inside \code{directory}.
Default is \code{"expr.csv"}. See details for the expected format.}

\item{\code{pheno_file}}{string. The name of the pheno data csv inside \code{directory}.
Default is \code{"pheno.csv"}. See details for the expected format.}

\item{\code{cohort}}{string in \code{c("train", "test")} or \code{NULL}. The cohort, train or test, to
prepare the data for. If NULL, the default, some functions will set it themselves
(e.g. \code{training_camp()} to \code{"train"}, \code{assess_2d()} to \code{"test"}), others will
throw an error.}

\item{\code{patient_id_col}}{string. The name of the column in the pheno data that holds
the patient identifiers. Default is \code{"patient_id"}.}

\item{\code{time_to_event_col}}{string. The name of the column in the pheno data that holds the
time-to-event values. Default is \code{"pfs_years"}.}

\item{\code{split_col_prefix}}{string. Column-name prefix for those columns holding splits into
train and test cohort. Some of these columns may be present in the pheno data already,
others will be added during the training process if required. Default is \code{"split_"},
i.e., split columns are named \code{"split_1"}, \code{"split_2"}, etc.}

\item{\code{event_col}}{string. The name of the column in the pheno data that holds
the event status encoded as 1 = occurrence, 0 = censoring. Default is
\code{"progression"}.}

\item{\code{benchmark_col}}{string or \code{NULL}. The name of the column in the pheno data that
holds the benchmark risk score (like the IPI). Default is \code{NULL}.}

\item{\code{gene_id_col}}{string. The name of the column in the expression data that holds
the gene identifiers. Default is \code{"gene_id"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The pheno csv file holds the samples as rows (with \emph{unique} sample ids in the
first column called \code{patient_id_col}), the variables as columns. The expr csv file
holds the genes as rows (with \emph{unique} gene ids in the first column called \code{gene_id_col}),
the samples as columns.
While the computational representation of both expression and pheno data will
change over the course of the pipeline, a DataSpec object will hold timeless
information on the data.
}

\subsection{Returns}{
A DataSpec object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-read"></a>}}
\if{latex}{\out{\hypertarget{method-Data-read}{}}}
\subsection{Method \code{read()}}{
Read expression data into the matrix \code{expr_mat} and pheno
data into the tibble \code{pheno_tbl}. Both will hold patients as rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
