% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_spec.R
\name{DataSpec}
\alias{DataSpec}
\title{Construct a DataSpec S3 object}
\usage{
DataSpec(
  name,
  directory,
  train_prop,
  pivot_time_cutoff = 2,
  expr_fname = "expr.csv",
  pheno_fname = "pheno.csv",
  patient_id_col = "patient_id",
  pfs_col = "pfs_years",
  progression_col = "progression",
  ipi_col = "ipi",
  ipi_feat_cols = c(age = "age", stage = "ann_arbor_stage", ldh_ratio = "ldh_ratio",
    performance_status = "ecog_performance_status", n_extranodal_sites =
    "n_extranodal_sites"),
  gene_id_col = "gene_id",
  split_col_prefix = "split_"
)
}
\arguments{
\item{name}{string. A telling name for the data set (e.g. \code{"Schmitz et al. 2018"}).}

\item{directory}{string. The directory where both expression and pheno csv
files lie.}

\item{train_prop}{numeric in (0, 1). The proportion of samples to draw for the
training cohort in every splitting run.}

\item{pivot_time_cutoff}{numeric or NULL. Pivotal time cutoff for the analysis and, explicitly,
for splitting the data into the train and test cohort: If a numeric in (0, 1), preserve
the the proportion of individuals below and above \code{time_cutoff} in both cohorts. Default
is \code{NULL}, i.e. no further constraints on splitting.}

\item{expr_fname}{string. The name of the expression csv file inside \code{directory}.
Default is \code{"expr.csv"}. See details for the expected format.}

\item{pheno_fname}{string. The name of the pheno data csv inside \code{directory}.
Default is \code{"pheno.csv"}. See details for the expected format.}

\item{patient_id_col}{string. The name of the column in the pheno data that holds
the patient identifiers. Default is \code{"patient_id"}.}

\item{pfs_col}{string. The name of the column in the pheno data that holds the
progression-free survival (PFS) values. Default is \code{"pfs_years"}.}

\item{progression_col}{string. The name of the column in the pheno data that holds
the progression status encoded as 1 = progression, 0 = no progression. Default is
\code{"progression"}.}

\item{ipi_col}{string. The name of the column in the pheno data that holds the
International Prognostic Index (IPI) values. Default is \code{"ipi"}.}

\item{ipi_feat_cols}{named character vector of length 5 or \code{NULL}. The names of the
\emph{five} features in the pheno data needed to compute the IPI. If NULL, no information
about the IPI features is provided, e.g. because they are not in the pheno data. For
the default, see Usage.}

\item{gene_id_col}{string. The name of the column in the expression data that holds
the gene identifiers. Default is \code{"gene_id"}.}

\item{split_col_prefix}{string. Column-name prefix for those columns holding splits into
train and test cohort. Some of these columns may be present in the pheno data already,
others will be added during the training process if required. Default is \code{"split_"},
i.e., split columns are named \code{"split_1"}, \code{"split_2"}, etc.}
}
\value{
A DataSpec object.
}
\description{
A DataSpec object specifies the location and format of the expression
and pheno data of a single data set. This enables reading and preparing the data.
}
\details{
The pheno csv file holds the samples as rows (with \emph{unique} sample ids in the
first column called \code{patient_id_col}), the variables as columns. The expr csv file
holds the genes as rows (with \emph{unique} gene ids in the first column called \code{gene_id_col}),
the samples as columns.
While the computational representation of both expression and pheno data will
change over the course of the pipeline, a DataSpec object will hold timeless
information on the data.
}
