% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{discretize_ipi_features}
\alias{discretize_ipi_features}
\title{Add discretized IPI features to a pheno_tbl}
\usage{
discretize_ipi_features(
  pheno_tbl,
  data_spec,
  cutoffs = c(age = 60, stage = 2, ldh_ratio = 1, performance_status = 1,
    n_extranodal_sites = 1)
)
}
\arguments{
\item{pheno_tbl}{tibble. The pheno data.}

\item{data_spec}{DataSpec. The data specification. See \code{DataSpec} for details.}

\item{cutoffs}{named numeric vector. The cutoff values for the IPI features. The names
of it must coincide with the names of \code{data_spec$ipi_feat_cols}. The default (see Usage)
takes the cutoffs from the original IPI paper \href{https://www.nejm.org/doi/full/10.1056/NEJM199309303291402}{ The International Non-Hodgkin's Lymphoma Prognostic Factors Project (1993): A Predictive Model for Aggressive Non-Hodgkin's Lymphoma }.}
}
\value{
The pheno tibble with the discretized IPI features appended.
}
\description{
Discretize the columns in a pheno tibble that belong to the features of
the International Prognostic Index (IPI) into binary indicator columns by comparing them
to cutoff values and append them to the pheno tibble.
}
