% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{discretize_tbl_cols}
\alias{discretize_tbl_cols}
\title{Discretize columns of a tibble and append the resulting columns}
\usage{
discretize_tbl_cols(tbl, col_names, cutoffs, gl = NULL)
}
\arguments{
\item{tbl}{tibble.}

\item{col_names}{character vector. Discretize the columns in \code{tbl} that have these
names.}

\item{cutoffs}{numeric vector of the same length as character vector. The cutoff values
corresponding to the columns in \code{col_names}.}

\item{gl}{character vector of the same length as \code{col_names} holding ">" and "<"
entries. Discretize via greater (\verb{> cutoffs}) or less (\verb{< cutoffs}). Default is \code{NULL},
in which case all columns are discretized via \verb{> cutoffs}.}
}
\value{
\code{tbl} with discretized columns as numeric columns appended.
}
\description{
Discretize the columns in a pheno tibble that belong to the features of
the International Prognostic Index (IPI) into binary indicator columns by comparing them
to cutoff values and append them to the pheno tibble.
}
