% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{download_and_read}
\alias{download_and_read}
\title{Download and read data into a tibble}
\usage{
download_and_read(data_dir, url, to_file, ...)
}
\arguments{
\item{data_dir}{string. The directory to store the downloaded file in.}

\item{url}{string. The url to download the file from.}

\item{to_file}{string. Dowloaded file will be stored as \code{to_file} inside
\code{data_dir}.}

\item{...}{Additional arguments to pass to the reader function. For example,
\code{readr::read_csv()} takes \code{col_types} as an argument.}
}
\value{
A tibble.
}
\description{
Download a csv, tsv or xlsx file from a url (if not already
downloaded), read it in and return it as a tibble.
}
