% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensures.R
\name{ensure_patients_match}
\alias{ensure_patients_match}
\title{Ensure patients in expression and pheno tibbles match during preprocessing}
\usage{
ensure_patients_match(
  expr_tbl,
  pheno_tbl,
  patient_id_col = "patient_id",
  gene_id_col = "gene_id",
  quiet = FALSE
)
}
\arguments{
\item{expr_tbl}{tibble. With genes as rows and patients as columns. The column
\code{gene_id_col} (see below) holds the gene identifiers, all other colum names
are the patient identifiers.}

\item{pheno_tbl}{tibble. The pheno data, with patients as rows and variables as
columns. The column \code{patient_id_col} (see below) holds the patient identifiers.}

\item{patient_id_col}{string. The name of the column in \code{pheno} that holds
the patient identifiers.}

\item{gene_id_col}{string. The name of the column in \code{expr} that holds the
gene identifiers.}

\item{quiet}{logical. Whether to print messages. Default is \code{FALSE}.}
}
\value{
A list with two tibbles named \code{expr} and \code{pheno}. Matching and sorted
expression and pheno data, i.e. the vectors of patient identifiers are
identical.
}
\description{
Ensure that the patients in the expression and pheno tibbles match
via subsetting and sorting, during preprocessing.
}
