% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{ensure_splits}
\alias{ensure_splits}
\title{Split a data set into train and test data multiple times}
\usage{
ensure_splits(pheno_tbl, data, model_list)
}
\arguments{
\item{pheno_tbl}{A tibble holding the pheno data (see \code{Data()}
for details).}

\item{data}{A \code{Data} object referring to \code{expr_tbl} and \code{pheno_tbl}.}

\item{model_list}{A list of \code{Model} objects. Ensure that every model
gets the splits it needs.}
}
\value{
\code{pheno_tbl} with all required splits.
}
\description{
Split a data set into train and test data sets by noting down
the affiliation to the train or test cohort in the pheno data for every split
Optionally, you can preserve the ratio of high- and low-risk patients. Ensure
that every model in a list of \code{Model}s gets enough splits.
}
