% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{error_rate}
\alias{error_rate}
\title{Error rate of a binary classifier}
\usage{
error_rate(y, y_hat)
}
\arguments{
\item{y}{A numeric vector with binary entries, the true outcomes.}

\item{y_hat}{A numeric vector with binary entries, the predicted outcomes.}
}
\value{
Numeric scalar, the error rate.
}
\description{
Error rate of a binary classifier
}
