% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit and store a model}
\usage{
fit(
  x,
  y,
  model,
  save_model_as,
  additional_model_par = list(),
  results_dir = "results",
  create_dir = TRUE,
  plots = TRUE
)
}
\arguments{
\item{x}{numeric matrix. The predictor matrix with samples in rows and their names
as row names.}

\item{y}{numeric matrix. The response matrix with samples in rows and their names
as row names.}

\item{model}{string. The name of the model to fit. Currently supported are
\itemize{
\item \code{cox_lasso_zerosum}, Cox proportional hazards with LASSO regularization and
zero-sum constraint,
\item \code{lasso_zerosum}, ordinary linear regression with LASSO regularization.
}}

\item{save_model_as}{character vector of length 2. The first element is the data-set
name from which data originates, the second element is the name of the very model
(usually a telling name in terms of \code{x}). See details.}

\item{additional_model_par}{list. Additional parameters to be passed to the model
fitting function, see \code{zeroSum::zeroSum} for details. Default is \code{list()}.}

\item{results_dir}{string. The directory holding results. Default is \code{"results"}.}

\item{create_dir}{logical. Whether to create the directory to store the model and
plots in if it does not exist. Default is \code{TRUE}.}

\item{plots}{logical. Whether to store a lambda-versus-deviance plot. Default is
\code{TRUE}.}
}
\value{
A \code{zeroSum.fit} object. The model fit.
}
\description{
Given a model and data in a model-specific form, fit this model
and store it
}
\details{
The model fit is stored in
\code{results_dir}/\code{save_model_as[1]}/\code{model}/\code{save_model_as[2]} with, e.g., \code{save_model_as  = c("schmitz", "with_ipi")}, meaning that \code{x} holds gene expression dara and the pheno
variabes needed to calculate the IPI score. Row names in \code{x} and \code{y} must match.
}
