% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{fitter_prototype}
\alias{fitter_prototype}
\title{Function interface of a fitter}
\usage{
fitter_prototype(x, y, val_error_fun, ...)
}
\arguments{
\item{x}{Named numeric matrix. Predictor matrix without \code{NA}s. Samples correspond to rows.
Discrete features are encoded as binary dummy variables.}

\item{y}{Named list with the response in thee formats:
\itemize{
\item \code{"bin"}, a named numeric one-column matrix, binary response,
\item \code{"cox"}, a named numeric two-column matrix, with columns \code{"time_to_event"} and \code{"event"}
(0 = censoring, 1 = event), the response in the Cox format,
\item \code{"true"}, a named numeric one-column matrix, true binary response.
}

The rownames of \code{y[["bin"]]} and \code{y[["true"]]} are a subset of the rownames of \code{x} and, in
general, do not coincide. Use \code{intersect_by_names()} to get equal rownames.}

\item{val_error_fun}{Function to calculate the error of validated predictions. For its interface,
see \code{\link[=val_error_fun_prototype]{val_error_fun_prototype()}}}

\item{...}{Further, fitter-specific hyperparameters.}
}
\value{
An S3 object with underlying class \code{list}, which we call \code{fit_obj}. The named list must
have the following element:
\itemize{
\item \code{"val_predict"}: a numeric one-column matrix with row names, the validated predictions of the
(picked) fitted model. The row names are a subset of the row names of \code{x}. The fitter may
tune hyperparameters and therefore fit multiple models. The \code{val_predict} attribute must contain
of the best validated model among them.
}

There must be a \code{predict()} method for the \code{fit_obj}. See \code{\link[=predict_method_prototype]{predict_method_prototype()}} for
its interface.
}
\description{
This function is the prototype of a fitter with the minimal requirements in its
parameters and return value to work as the \code{fitter} attribute of a \code{Model} object.
}
