% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xy.R
\name{generate_predictor}
\alias{generate_predictor}
\title{Generate the predictor matrix in a model-specific way}
\usage{
generate_predictor(expr_mat, pheno_tbl, data, model)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression data, with patients as rows
and genes as columns.}

\item{pheno_tbl}{tibble. The pheno data, with patients as rows and variables as
columns.}

\item{data}{Data S3 object. Specifications on the data (\code{expr_mat} and
\code{model}). See the the constructor \code{Data()} for details.}

\item{model}{Model S3 object. Specifications on the model to prepare for.
See the the constructor \code{Model$new()} for details.}
}
\value{
A named numeric matrix with patients as rows and variables as columns. No
\code{NA}s in it.
}
\description{
Generate the numeric predictor matrix from the expression and
(possibly) pheno data for a certain model
}
