% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestor.R
\name{greedy_nestor}
\alias{greedy_nestor}
\title{Nest an existing, tuned model together with more features into another
model and tune the latter}
\usage{
greedy_nestor(x, y, val_error_fun, model1, fitter2, hyperparams2)
}
\arguments{
\item{x}{named numeric matrix (samples x features). Features only meant for
the late model are exactly those matching \code{x}'s \code{li_var_suffix} attribute.}

\item{y}{Named list with entries
\itemize{
\item \code{"bin"}, a named numeric one-column matrix, binary response to be used
for training,
\item \code{"cox"}, a named numeric two-column matrix to be used for training, time to
event and event (0 = censoring, 1 = event) in first and second column, respectively.
\item \code{"true"}, a named numeric one-column matrix, binary response to be used for
calculating the CV error.
}}

\item{val_error_fun}{Function used to calculate the error of independently
validated predictions. Must take two numeric vector of equal length:
\code{y} and \code{y_hat}, the true and predicted outcomes, respectively, and
return a numeric scalar; the lower, the better the model. See
\code{\link[=error_rate]{error_rate()}} or \code{\link[=neg_roc_auc]{neg_roc_auc()}} for examples.}

\item{model1}{\code{Model} R6 object. The early model trained on the expression
data, with the \code{fits} attribute set at least in its stored version, i.e., the
early model is already there.}

\item{fitter2}{A \emph{patroklos-compliant fitter with CV tuning} (see
README for more details).}

\item{hyperparams2}{A named list with hyperparameters for the late model.}
}
\value{
A \code{nested_fit} S3 object.
}
\description{
Use the validated predictions of an existing model that only
takes the epression part of the data as input features, and feed them
together with the remaining features into another model. Fit and tune the
second, late model.
}
