% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{imputer_prototype}
\alias{imputer_prototype}
\title{Function interface of an imputer}
\usage{
imputer_prototype(x)
}
\arguments{
\item{x}{Numeric matrix. Predictor matrix. Samples correspond to rows. Discrete features are
encoded as binary dummy variables. May contain \code{NA}s.}
}
\value{
Numeric matrix. Predictor matrix of same shape as \code{x} with non-\code{NA}s untouched.
}
\description{
Function interface of an imputer
}
