% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_cv.R
\name{integrated_cv}
\alias{integrated_cv}
\title{Build a fitter with integrated cross-validation from a fitter}
\usage{
integrated_cv(
  fitter,
  metric = c("accuracy", "roc_auc", "binomial_log_likelihood")
)
}
\arguments{
\item{fitter}{A patroklos-compliant fitter.}

\item{metric}{A character string or a function. If a character string, it
must be one of "accuracy", "roc_auc", or "binomial_log_likelihood". If a
function, it must take two arguments: \code{y} and \code{y_hat}, and return a numeric
scalar. The returned fitter will use it to calculate the goodness of validated
predictions.}
}
\value{
A patroklos-compliant fitter with integrated cross-validation.
}
\description{
This function operator takes a patroklos-compliant fitter and
builds a patroklos-compliant fitter with integrated cross validation from
it.
}
