% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{mean_impute}
\alias{mean_impute}
\title{Impute missing values in a matrix by column means}
\usage{
mean_impute(x)
}
\arguments{
\item{x}{numeric matrix. Sample as rows, variables as columns. Impute for
\code{NA}s in the columns.}
}
\description{
For categorical variables (which we expect to be dichotomized
already), we also simply compute the mean, which is the (marginal) Bernoullie
probability to be in the positive class (1).
}
