% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mirror_path}
\alias{mirror_path}
\title{Mirror a file path}
\usage{
mirror_path(filepath, mirror)
}
\arguments{
\item{filepath}{string. The file path to mirror.}

\item{mirror}{character vector of length 2. Replace mirror[1] by mirror[2] in
filepath.}
}
\value{
A string. The mirrored file path.
}
\description{
Mirror a file path
}
