% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R
\name{mock_data_from_existing}
\alias{mock_data_from_existing}
\title{Subset existing data to mock data}
\usage{
mock_data_from_existing(
  directory,
  n_samples = 50,
  n_genes = 10,
  expr_fname = "expr.csv",
  pheno_fname = "pheno.csv",
  patient_id_col = "patient_id",
  gene_id_col = "gene_id",
  save = TRUE,
  save_suffix = "mock"
)
}
\arguments{
\item{directory}{string. The directory where both expression and pheno csv
files lie.}

\item{n_samples}{integer. The number of samples to randomly pick from the
initial data.}

\item{n_genes}{integer. The number of genes to randomly pick from the
initial data.}

\item{expr_fname}{string. The name of the expression csv file inside \code{directory}.
See details of \code{read} for the expected format. Default is \code{"expr.csv"}.}

\item{pheno_fname}{string. The name of the pheno data csv inside \code{directory}.
See details of \code{read} for the expected format. Default is \code{"pheno.csv"}.}

\item{patient_id_col}{string. The name of the column in the pheno data csv
file that holds the patient identifiers. Default is \code{"patient_id"}.}

\item{gene_id_col}{string. The name of the column in the expression data csv
file that holds the gene identifiers. Default is \code{"gene_id"}.}

\item{save}{logical. Whether to save the subset as csv files. Default is
\code{TRUE}.}

\item{save_suffix}{string. The suffix to append to the file names of the
subset csv files. Default is \code{"mock"}.}
}
\value{
A list with two tibbles, \code{expr} and \code{pheno}, the subset expression
and pheno data. If \code{save} is \code{TRUE}, the subset data will be saved as csv
files in the format desired by \code{read}.
}
\description{
Read in the expression and pheno data of an existing dataset,
randomly pick a subset of samples and genes, and return and possibly save
the subset as mock data.
}
