% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model}
\alias{Model}
\title{An R6 class for a model}
\description{
A Model specifies how a model looks like, prepares the data, fit_obj it,
stores it and predicts from it.
}
\seealso{
\code{\link[=training_camp]{training_camp()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A telling name for the model.}

\item{\code{directory}}{Store/find the models in this directory.}

\item{\code{fitter}}{The model fitting function to be used.}

\item{\code{time_cutoffs}}{Threshold and censor the outcome accordingly.}

\item{\code{val_error_fun}}{Calculate the error of independently validated predictions.}

\item{\code{hyperparams}}{Optional arguments passed to \code{fitter}.}

\item{\code{include_from_continuous_pheno}}{The names of the continuous variables in the
pheno data (to be) included in the predictor matrix.}

\item{\code{include_from_discrete_pheno}}{The names of the discrete variables in the
pheno data (to be) included in the predictor matrix.}

\item{\code{include_expr}}{Whether to include the expression data in the predictor matrix.}

\item{\code{li_var_suffix}}{Append this to the names of features from the pheno data
when adding them to the predictor matrix.}

\item{\code{create_directory}}{Whether to create \code{directory} if it does not exist, yet.}

\item{\code{file}}{Store this Model object under this name in \code{directory}.}

\item{\code{fit_obj}}{A list holding fit_obj (something returned by a fitter like
a \code{ptk_zerosum} S3 object).}

\item{\code{continuous_output}}{Whether the output of the model is continous or binary.}

\item{\code{combine_n_max_categorical_features}}{Maximum number of categorical features
to combine}

\item{\code{combined_feature_min_positive_ratio}}{Minimum ratio of positive
observations in a combined (categorical) feature}

\item{\code{enable_imputation}}{Overrides the \code{imputer} attribute of the \code{Data} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Model-new}{\code{Model$new()}}
\item \href{#method-Model-fit}{\code{Model$fit()}}
\item \href{#method-Model-predict}{\code{Model$predict()}}
\item \href{#method-Model-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-new"></a>}}
\if{latex}{\out{\hypertarget{method-Model-new}{}}}
\subsection{Method \code{new()}}{
Create a new Model instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(
  name,
  fitter,
  directory,
  time_cutoffs,
  val_error_fun,
  hyperparams = NULL,
  include_from_continuous_pheno = NULL,
  include_from_discrete_pheno = NULL,
  include_expr = TRUE,
  li_var_suffix = "++",
  create_directory = TRUE,
  file = "model.rds",
  continuous_output = NULL,
  combine_n_max_categorical_features = 1L,
  combined_feature_min_positive_ratio = 0.04,
  enable_imputation = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string. A telling name for the model.}

\item{\code{fitter}}{function. The model fitting function to be used. Must take \code{x} and
\code{y} as first two positional arguments. Further arguments can be passed via
\code{hyperparams} (below). Its return value must be an S3 object with a \code{plot()}
method, and (ideally, for assessment) with a \code{predict()} method. Default is \code{NULL}.}

\item{\code{directory}}{string. The directory to store the models in. For every value in
\code{time_cutoffs}, find the corresponding model in a subdirectory named after this value.}

\item{\code{time_cutoffs}}{numeric vector. This governs how the provided
response looks like. For every value of \code{time_cutoffs}, specify a model
on the following response data:
\itemize{
\item For Cox response, censor all patients where the event ouccured after
\code{time_cutoffs} at this value and train the specified model.
\item For binary response, binarize the outcome depending on whether it
occured before or after this value.
}

When fitting models to the data, you find every of the
\code{length(time_cutoffs)} models in a subdirectory of \code{directory} named
after the respective value in \code{time_cutoffs}.}

\item{\code{val_error_fun}}{Function used to calculate the error of independently
validated predictions. Must take two numeric vector of equal length:
\code{y} and \code{y_hat}, the true and predicted outcomes, respectively, and
return a numeric scalar; the lower, the better the model. See
\code{\link[=error_rate]{error_rate()}} or \code{\link[=neg_roc_auc]{neg_roc_auc()}} for examples.}

\item{\code{hyperparams}}{list. Optional arguments passed to \code{fitter}, e.g. alpha
in case of an elastic net. Default is \code{list()}, i.e., no arguments other than \code{x}, \code{y}
passed to \code{fitter}.}

\item{\code{include_from_continuous_pheno}}{vector of strings. The names of the
\emph{continuous} variables in the pheno data (to be) included in the predictor matrix. The
values will be coerced to numeric. Default is \code{NULL}, which means no continuous pheno
variables are or will be included.}

\item{\code{include_from_discrete_pheno}}{vector of strings. The names of the \emph{discrete}
variables in the pheno data (to be) included in the predictor matrix. A discrete
variable with n levels will be converted to n-1 binary variables. Default is \code{NULL},
which means no discrete pheno variables are or will be included.}

\item{\code{include_expr}}{logical. Whether to include the expression data in the predictor
matrix.}

\item{\code{li_var_suffix}}{string. Append this to the names of features from the pheno
data when adding them to the predictor matrix. Default is \code{"++"}.}

\item{\code{create_directory}}{logical. Whether to create \code{directory} if it does not exist, yet.
Default is \code{TRUE}.}

\item{\code{file}}{string. The name of the model-fit_obj file inside \code{directory}.
Default is \code{"fit_obj.rds"}.}

\item{\code{continuous_output}}{logical or NULL. Whether the output of the model
is continous (like the conditional probabilities of a logistic regression) or
binary (like with a random forest). This piece of information is helpful
when it comes to assessing a model (Ass2d$assess_center() uses it, e.g.).}

\item{\code{combine_n_max_categorical_features}}{integer. Maximum number of categorical features
to combine in predicting features.}

\item{\code{combined_feature_min_positive_ratio}}{numeric. Minimum ratio of positive
observations in a combined (categorical) feature. This attribute together
with \code{combine_n_max_categorical_features} governs which combinatorial
features the predicitor matrix will contain.}

\item{\code{enable_imputation}}{logical. If \code{FALSE}, it overrides the \code{imputer} attribute of
the \code{Data} objectl and we do not impute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Model} R6 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Model-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model to a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$fit(data, quiet = FALSE, msg_prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Read it in if needed.}

\item{\code{quiet}}{logical. Whether to suppress messages. Default is \code{FALSE}.}

\item{\code{msg_prefix}}{string. Prefix for messages. Default is \code{""}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{Model} object itself with the \code{fit_obj} attribute set
to the object tuned over \code{time_cutoffs}, \code{combine_n_max_categorical_features}
and \code{hyperparams} with \code{fitter} at the core.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Model-predict}{}}}
\subsection{Method \code{predict()}}{
Predict for a data set. Additonally return the true values
of the response and, if the \code{benchmark_col} attribute of the \code{Data}
object is not \code{NULL}, the values of the benchmark.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$predict(data, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Specifications on the data. Read it in if
needed.}

\item{\code{quiet}}{logical. Whether to suppress messages. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of three named numeric vectors named \code{"predictions"},
\code{"actual"}, and \code{"benchmark"}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
