% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{ModelSpec}
\alias{ModelSpec}
\title{Construct a ModelSpec S3 object}
\usage{
ModelSpec(
  fitter,
  optional_fitter_args = NULL,
  response_type = NULL,
  include_from_continuous_pheno = NULL,
  include_from_discrete_pheno = NULL,
  pfs_leq = NULL,
  save_dir = ".",
  create_save_dir = TRUE,
  plot_fname = "training_error.pdf",
  fit_fname = "fit_obj.rds"
)
}
\arguments{
\item{fitter}{function. The model fitting function to be used. Must take \code{x} and
\code{y} as first two positional arguments. Further arguments can be passed via
\code{optional_fitter_args} (below). Its return value must be an S3 object with a \code{plot()}
method, and (ideally, for assessment) with a \code{predict()} method. Default is \code{NULL}.}

\item{optional_fitter_args}{list. Optional arguments passed to \code{fitter}, e.g. alpha
in case of an elastic net. Default is \code{list()}, i.e., no arguments other than \code{x}, \code{y}
passed to \code{fitter}.}

\item{response_type}{string. The type of response to be used. One of \code{"binary"} or
\code{"survival_censored"}. Default is \code{NULL}.}

\item{include_from_continuous_pheno}{vector of strings. The names of the
\emph{continuous} variables in the pheno data (to be) included in the predictor matrix. The
values will be coerced to numeric. Default is \code{NULL}, which means no continuous pheno
variables are or will be included.}

\item{include_from_discrete_pheno}{vector of strings. The names of the \emph{discrete}
variables in the pheno data (to be) included in the predictor matrix. A discrete
variable with n levels will be converted to n-1 binary variables. Default is \code{NULL},
which means no discrete pheno variables are or will be included.}

\item{pfs_leq}{numeric. Only used if \code{response_type == "binary"}. The value of
progression-free survival (PFS) below which samples are considered high-risk. Default
is \code{NULL}.}

\item{save_dir}{string. The directory in which to store the model in. Default is \code{"."},
the current working directory.}

\item{create_save_dir}{logical. Whether to create \code{save_dir} if it does not exist, yet.
Default is \code{TRUE}.}

\item{plot_fname}{string. Store the plot resulting from \code{plot(fit_obj)} in \code{save_dir}
under this name. Default is \code{"training_error.pdf"}.}

\item{fit_fname}{string. The name of the model-fit file inside \code{save_dir}.
Default is \code{"fit_obj.rds"}.}
}
\value{
A ModelSpec S3 object.
}
\description{
A ModelSpec object holds all the tools and information needed to fit and
store a model. It is used as an argument to \code{fit()} and \code{prepare_and_fit()}. Its base
object is a list containing:
}
