% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{multitune_output_prototype}
\alias{multitune_output_prototype}
\title{Function interface for the return value of \code{\link[=multitune]{multitune()}}}
\usage{
multitune_output_prototype(x, y, val_error_fun, ...)
}
\arguments{
\item{x}{Named numeric matrix. Predictor matrix without \code{NA}s. Samples correspond to rows.
Discrete features are encoded as binary dummy variables.}

\item{y}{Named list with the response in thee formats:
\itemize{
\item \code{"bin"}, a named numeric one-column matrix, binary response,
\item \code{"cox"}, a named numeric two-column matrix, with columns \code{"time_to_event"} and \code{"event"}
(0 = censoring, 1 = event), the response in the Cox format,
\item \code{"true"}, a named numeric one-column matrix, true binary response.
}

The rownames of \code{y[["bin"]]} and \code{y[["true"]]} are a subset of the rownames of \code{x} and, in
general, do not coincide. Use \code{intersect_by_names()} to get equal rownames.}

\item{val_error_fun}{Function to calculate the error of validated predictions. For its interface,
see \code{\link[=val_error_fun_prototype]{val_error_fun_prototype()}}}

\item{...}{atomic vectors. Every vector corresponds to a hyperparameter and holds candidate
values for it. For every combination of hyperparameters, fit a model by calling the \code{fitter}
parameter of \code{multitune()}.}
}
\value{
A \code{multitune_obj} S3 object with underlying class \code{list}. It fulfills the requirements
of \code{fitter_prototype()} and, in addition, it has the following elements:
\itemize{
\item \code{val_predict_list}: a list of row-named one-column matrices, the validated predictions of the
fitted models.
\item \code{lambda}: a character vector, the hyperparameter combinations as a string.
\item \code{lambda_min_index}: an integer, the index of the hyperparameter combination of the model with
the lowest validation error.
\item \code{lambda_min}: a character, the hyperparameter combination of the model with the lowest
validation error.
\item \code{val_error}: a numeric vector, the validation errors of the fitted models.
\item \code{min_error}: a numeric scalar, the validation error of the model with the lowest validation
error.
\item \code{fit_obj_list}: a list of \code{fit_obj}s of the fitted models. If \code{select = TRUE} in
\code{\link[=multitune]{multitune()}}, all but the model with the lowest validation error are \code{NA}.
}
}
\description{
Function interface for the return value of \code{\link[=multitune]{multitune()}}
}
