% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cv.R
\name{neg_binomial_log_likelihood}
\alias{neg_binomial_log_likelihood}
\title{Negative binomial log-likelihood}
\usage{
neg_binomial_log_likelihood(y, y_hat)
}
\arguments{
\item{y}{A numeric vector with binary entries, the true outcomes.}

\item{y_hat}{A numeric vector with continuous entries, the predicted outcomes.}
}
\value{
Numeric scalar, the negative binomial log-likelihood.
}
\description{
Negative binomial log-likelihood
}
