% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cv.R
\name{neg_prec_with_prev_greater}
\alias{neg_prec_with_prev_greater}
\title{Minimal negative precision for thresholds with a minimal prevalence}
\usage{
neg_prec_with_prev_greater(min_prev)
}
\arguments{
\item{min_prev}{A numeric scalar, the minimal prevalence.}
}
\value{
A function that takes two numeric vectors \code{y} and \code{y_hat} and returns
the minimal precision over those thresholds yielding a prevalence of at least
\code{min_prev}.
}
\description{
A function factory.
}
