% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{neg_roc_auc}
\alias{neg_roc_auc}
\title{Negative ROC AUC}
\usage{
neg_roc_auc(y, y_hat)
}
\arguments{
\item{y}{A numeric vector with binary entries, the true outcomes.}

\item{y_hat}{A numeric vector with continuous entries, the predicted outcomes.}
}
\value{
Numeric scalar, the negative ROC AUC.
}
\description{
Negative ROC AUC
}
