% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cv.R
\name{nested_fit}
\alias{nested_fit}
\title{Construct a nested_fit S3 object}
\usage{
nested_fit(model1, model2, search_grid, best_hyperparams)
}
\arguments{
\item{model1}{An S3 object with a \code{predict} method. The early model.}

\item{model2}{An S3 object with a \code{predict} method. The late model.}

\item{search_grid}{A data frame with the hyperparamters for the early
(\code{lambda}) and the late model, i.e., one row corresponds to one trained
model or to one combinations of hyperparameters. Plus an extra column for the
cross-validated or out-of-bag (OOB) accuracy.}

\item{best_hyperparams}{A named list with the best hyperparameters.}
}
\value{
An S3 object with class `nested_fit'.
}
\description{
Construct a nested_fit object that holds the early and late model
plus their hyperparameters.
}
\details{
Nest two models. The first model takes part of the features (those
not matching pattern) to make a prediction that we in turn feed together with
the rest of the features to the second model. This gives us a final prediction.
Both models must be S3 objects with a \code{predict} method.
}
