% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R, R/nestor.R
\name{non_zero_coefs}
\alias{non_zero_coefs}
\alias{non_zero_coefs.ptk_zerosum}
\alias{non_zero_coefs.nested_fit}
\title{Get features with non-zero coefficients}
\usage{
non_zero_coefs(fit_obj)

\method{non_zero_coefs}{ptk_zerosum}(fit_obj)

\method{non_zero_coefs}{nested_fit}(fit_obj)
}
\arguments{
\item{fit_obj}{S3 fit object.}
}
\value{
One-column matrix with non-zero coefficients and feature names as row names.
}
\description{
Get features with non-zero coefficients
}
\details{
Throws an error if method for a class is not implemented (e.g. for \code{ptk_ranger}
because "non-zero coefficients" do not make sense with random forests).

Refers to the late model.
}
