% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_plot_spec.R
\name{PerfPlotSpec}
\alias{PerfPlotSpec}
\title{Create a PerfPlotSpec object}
\usage{
PerfPlotSpec(
  fname,
  x_metric,
  y_metric,
  pfs_leq = NULL,
  lambda = "lambda.min",
  benchmark = "ipi",
  fellow_csv = TRUE,
  scores_plot = TRUE,
  show_plots = FALSE,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  xlim = NULL,
  ylim = NULL,
  smooth_method = NULL,
  smooth_benchmark = FALSE,
  alpha = 0.5,
  width = 7,
  height = 4,
  colors = NULL,
  units = "in"
)
}
\arguments{
\item{fname}{string. The name of the file to save the plot to.}

\item{x_metric}{string. The name of the performance measure to be plotted on the x-axis.
All measures that can be passed to the \code{x.measure} parameter of \code{\link[ROCR:performance]{ROCR::performance()}} are
valid.}

\item{y_metric}{string. The name of the performance measure to be plotted on the y-axis.
All measures that can be passed to the \code{measure} parameter of \code{\link[ROCR:performance]{ROCR::performance()}} are
valid.}

\item{pfs_leq}{numeric. Progression-free survival threshold that divides samples into
high-risk (PFS < \code{pfs_leq}) and low-risk (PFS >= \code{pfs_leq}) group. Model performance
will be measured in terms of how well it can separate these two groups as a binary
classifier. Default is \code{2.0}.}

\item{lambda}{numeric or string. The lambda regularization parameter of the model to
evaluate. See \code{\link[=prepare_and_predict]{prepare_and_predict()}} for details. Default is \code{"lambda.min"}.}

\item{benchmark}{string. Column in the test pheno holding the numeric benchmark values.
Default is \code{"ipi"} (international prognostic index for DLBCL).}

\item{fellow_csv}{logical. If passed to \code{\link[=assess_multiple_models]{assess_multiple_models()}}, whether to also create a
csv file for every model-data pair. Default is \code{TRUE}.}

\item{scores_plot}{logical. Display the ordered scores output by the model in a scatter
plot. Default is \code{TRUE}.}

\item{show_plots}{logical. Whether to show the plots after creation in an interactive
graphics device. Default is \code{FALSE}.}

\item{title}{string. The title of the plot. Default is \code{NULL}.}

\item{x_lab, y_lab}{string. Axis labels. Default is \code{x_metric} and \code{y_metric}, respectively.}

\item{xlim, ylim}{numeric vector of length 2 or \code{NULL}. The limits for both axes. Default is
\code{NULL}, no limits.}

\item{smooth_method}{string or function. Smooth method to plot an additional smoothed graph.
If \code{NULL}, no smoothing. Else we pass \code{smooth_method} as the \code{method} parameter to
\code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}. Default is \code{"loess"}.}

\item{smooth_benchmark}{logical. Whether to also smooth the benchmark data. Default is
\code{FALSE}.}

\item{alpha}{numeric in [0, 1]. The alpha value for the points and lines in the
plot.}

\item{width}{numeric. The width of the plot in \code{units}. Default is \code{7}.}

\item{height}{numeric. The height of the plot in \code{units}. Default is \code{4}.}

\item{colors}{character vector. The colors to be used for the different models.
Default is \code{NULL}, which means that the default colors of \code{ggplot2} will be used.}

\item{units}{string. The units of \code{width} and \code{height}. Default is \code{"in"} (inches).}
}
\value{
A PlotSpec S3 object.
}
\description{
A PlotSpec object holds all the info on how to create a model performance
plot and what to do with it. It is tailored for \code{\link[=assess_model]{assess_model()}}. Its base object is a
list containing:
}
