% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestor.R
\name{predict.nested_fit}
\alias{predict.nested_fit}
\title{Predict method for nested_fit objects}
\usage{
\method{predict}{nested_fit}(object, newx, ...)
}
\arguments{
\item{object}{An S3 object with class \code{nested_fit}. The two-stage nested
model.}

\item{newx}{A numeric matrix with the same number of columns and column names
as the matrix used to fit the model. Rows correspond to samples, columns to
features.}

\item{...}{Other arguments for the normal predict function if the fit is not
a nested_fit object.}
}
\value{
A numeric vector with the prediction for every sample.
}
\description{
Predict method for nested_fit objects
}
