% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{predict.ptk_ranger}
\alias{predict.ptk_ranger}
\title{Wrap \code{\link[ranger:predict.ranger]{ranger::predict.ranger()}} into a patroklos-compliant predict
function}
\usage{
\method{predict}{ptk_ranger}(object, newx, ...)
}
\arguments{
\item{object}{A ptk_ranger S3 object.}

\item{newx}{Predictor data as a named numeric matrix. Samples correspond to
rows, an attribute \code{li_var_suffix} is expected to be present.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A named numeric vector of predictions.
}
\description{
\code{Model$predict()} uses this function as a predict method for
the \code{predict()} generic.
}
\details{
The interface of this function is the interface of a patroklos-compliant
predict method.
}
