% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{predict.ptk_zerosum}
\alias{predict.ptk_zerosum}
\title{Wrap \code{\link[zeroSum:predict.zeroSum]{zeroSum::predict.zeroSum()}} into a patroklos-compliant predict function}
\usage{
\method{predict}{ptk_zerosum}(object, newx, ...)
}
\arguments{
\item{object}{A ptk_patroklos S3 object.}

\item{newx}{Predictor data as a named numeric matrix. Samples correspond to
rows, an attribute \code{li_var_suffix} is expected to be present.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A numeric vector with the prediction for every sample. If the
\code{binarize_predictions} attribute of the \code{ptk_zerosum} object is NULL, return
the linear predictor beta_0 + beta^T x. Otherwise, return the binary class
labels by thresholding the logistic transformation of the linear predictor
via \code{binarize_predictions}.
}
\description{
\code{Model$predict()} uses this function as a predict method for
the \code{predict()} generic.
}
