% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{predict_method_prototype}
\alias{predict_method_prototype}
\title{Function interface of the S3 generic \code{predict()} for a \code{fit_obj}}
\usage{
predict_method_prototype(object, newx, ...)
}
\arguments{
\item{object}{The fitted S3 object.}

\item{newx}{Numeric matrix. The predictor matrix without \code{NA}s. Samples correspond to rows.
Discrete features are encoded as binary dummy variables. Column names are the same as in the
predictor matrix used for fitting.}

\item{...}{Further parameters specific for the S3 class of \code{object}. Ignored by
\code{Model$predict()}.}
}
\value{
A numeric one-column matrix with as many rows as \code{newx}, the predictions of the (picked)
fitted model for the samples in \code{newx}. The fitter that outputs \code{object} may fit multiple models
and store them in \code{object}, but the \code{predict()} method must return the predictions of the best
model among them.
}
\description{
This function is the prototype of a \code{predict()} S3 method for the S3 object
\code{fit_obj} of a \code{Model} and lays out the minimal requirements for its parameters and return value.
}
