% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data for model fitting and predicting}
\usage{
prepare(
  expr_mat,
  pheno_tbl,
  model,
  include_from_continuous_pheno = NULL,
  include_from_discrete_pheno = NULL,
  patient_id_col = "patient_id",
  pfs_col = "pfs_years",
  progression_col = "progression",
  pfs_leq = 2
)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression data, with patients as rows
and genes as columns. The row names must be unique patient identifiers.}

\item{pheno_tbl}{tibble. The pheno data, with patients as rows and variables as
columns.}

\item{model}{string. For which model to prepare. One of
\itemize{
\item \code{"cox_lasso_zerosum"} (Cox proportional hazards with LASSO reularization
and zero-sum constraint)
\item \code{"lasso_zerosum"} (ordinary linear regression with LASSO regularization)
}}

\item{include_from_continuous_pheno}{vector of strings. The names of the
\emph{continuous} variables in the pheno data file to be included in the
predictor matrix. The values will be coerced to numeric. Default is \code{NULL},
which means no continuous pheno variables will be included.}

\item{include_from_discrete_pheno}{vector of strings. The names of the
\emph{discrete} variables in the pheno data file to be included in the predictor
matrix. A discrete variable with n levels will be converted to n-1 binary
variables. Default is \code{NULL}, which means no discrete pheno variables will
be included.}

\item{patient_id_col}{string. The name of the column in \code{pheno_tbl}
that holds the patient identifiers. Default is \code{"patient_id"}.}

\item{pfs_col}{string. The name of the column in \code{pheno_tbl} that holds the
progression-free survival (PFS) values. Default is \code{"pfs_years"}.}

\item{progression_col}{string. The name of the column in \code{pheno_tbl} that
holds the progression status encoded as 1 = progression, 0 = no progression.
Default is \code{"progression"}.}

\item{pfs_leq}{numeric. Only necessary if model discretizes response (PFS). The value
of progression-free survival (PFS) above which samples are considered high-risk.
Default is 2.0.}
}
\value{
A list with two numeric matrices, \code{x} and \code{y}. \code{x} is the predictor matrix,
\code{y} is the response matrix, both ready to for model fitting or predicting.
}
\description{
Provided expression matrix and pheno tibble, generate the
predictor and respondent matirx model-specifically and check.
}
