% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_fit.R
\name{prepare_and_fit}
\alias{prepare_and_fit}
\title{Prepare data, fit and store models}
\usage{
prepare_and_fit(expr_mat, pheno_tbl, data_spec, model_spec)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix with genes in rows and samples
in columns.}

\item{pheno_tbl}{tibble. The pheno data with samples in rows and variables in columns.}

\item{data_spec}{DataSpec S3 object. Specifications on the data. See the the
constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec S3 object. Specifications on the model. See
the the constructor \code{ModelSpec()} for details.}
}
\value{
A list of fit objects as returned by the \code{fit()} method of \code{model_spec}.
}
\description{
Given an expression matrix and a pheno tibble, prepare the data for a
list of models and fit these models to the data.
}
