% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_predict.R
\name{prepare_and_predict}
\alias{prepare_and_predict}
\title{Prepare data and predict with a loaded model}
\usage{
prepare_and_predict(expr_mat, pheno_tbl, data_spec, model_spec, lambda)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix with genes in rows and samples
in columns.}

\item{pheno_tbl}{tibble. The pheno data with samples in rows and variables in columns.}

\item{data_spec}{DataSpec S3 object. Specifications on the data. See the the
constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec S3 object. Specifications on the model. See the the
constructor \code{ModelSpec()} for details.}

\item{lambda}{string or numeric. The lambda regularization parameter of the model
to predict with. Technically, we will pass it to the \code{s} parameter of \code{predict()}
method of the object returned by the \code{fitter} attribute of the \code{ModelSpec} object.
See, e.g., \code{\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet()}} or \code{\link[zeroSum:predict.zeroSum]{zeroSum::predict.zeroSum()}}.}
}
\value{
A list with two numeric vectors:
\itemize{
\item \code{"predictions"}: predicted scores,
\item "actual": actual, \emph{according to \code{model_spec$pfs_leq} discretized} response
}
}
\description{
Given an expression matrix and a pheno tibble, prepare the data for a
certain model and predict with this model.
}
