% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{prepend_to_directory}
\alias{prepend_to_directory}
\title{To every \code{Model} in a list of \code{Model}s, prepend a fixed directory to the
\code{directory} attribute}
\usage{
prepend_to_directory(model_list, prefix)
}
\arguments{
\item{model_list}{list of \code{Model}s.}

\item{prefix}{character Prepend this directory to the \code{directory} attribute of each
\code{Model} in \code{model_list}.}
}
\value{
list of modified \code{Model}s.
}
\description{
This function supports the paradigm of specifying the \code{Model}s for
for \emph{all} data sets at one place and them quickly adopting for a specific data
set. We modify the Model objects \emph{in place}.
}
