% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass_scalar_helpers.R
\name{prepend_to_filename}
\alias{prepend_to_filename}
\title{Prepend a prefix to the file attribute of a list of AssScalar objects}
\usage{
prepend_to_filename(ass_scalar_list, prefix)
}
\arguments{
\item{ass_scalar_list}{list of AssScalar objects.}

\item{prefix}{character Prepend this prefix to the \code{file} attribute of each
element in \code{ass_scalar_list}.}
}
\value{
list of modified \code{AssScalar}s.
}
\description{
Modify the AssScalar objects \emph{in place}. This function supports
the paradigm of specifying the \code{Model}s for
for \emph{all} data sets at one place and them quickly adopting for a specific data
set.
}
