% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_spec.R
\name{prepend_to_save_dir}
\alias{prepend_to_save_dir}
\title{To every \code{ModelSpec} in a list of \code{ModelSpec}s, prepend a fixed directory to the
\code{save_dir} attribute}
\usage{
prepend_to_save_dir(model_spec_list, dir)
}
\arguments{
\item{model_spec_list}{list of \code{ModelSpec}s.}

\item{dir}{string. The directory to prepend to the \code{save_dir} attribute of all
\code{ModelSpec}s in \code{model_spec_list}.}
}
\value{
A list of \code{ModelSpec}s, with \code{dir} prepended to the \code{save_dir} attribute.
}
\description{
Often one specifies the models in general, for all data sets. If you fit the
models to a specific data set (say \code{"mock"}), you might want to prepend a fixed directory
like \code{"results/mock"} to the \code{save_dir} attribute of all \code{ModelSpec}s in the list.
}
