% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{projection_on_feature}
\alias{projection_on_feature}
\title{Project on a feature}
\usage{
projection_on_feature(x, y, val_error_fun, feature, ...)
}
\arguments{
\item{x}{data as a numeric matrix object (rows=samples). The zero-sum
regression requires data on the log scale, i.e. x should be
log-transformed data.}

\item{y}{Named list with entries
\itemize{
\item \code{"bin"}, a named numeric one-column matrix, binary response to be used
for training,
\item \code{"cox"}, a named numeric two-column matrix to be used for training, time to
event and event (0 = censoring, 1 = event) in first and second column, respectively.
\item \code{"true"}, a named numeric one-column matrix, binary response to be used for
calculating the CV error.
}}

\item{val_error_fun}{Function used to calculate the error of independently
validated predictions. Must take two numeric vector of equal length:
\code{y} and \code{y_hat}, the true and predicted outcomes, respectively, and
return a numeric scalar; the lower, the better the model. See
\code{\link[=error_rate]{error_rate()}} or \code{\link[=neg_roc_auc]{neg_roc_auc()}} for examples.}

\item{feature}{character. Project on this feature.}

\item{...}{can be used for adjusting internal parameters}
}
\value{
A \code{ptk_zerosum} S3 object with the \code{val_predict} attribute set to the predictions on
the "train" cohort. The benchmark usually has not been trained on this train cohort such that
\code{val_predict} are predictions on independent data.
}
\description{
Trivially wraps \code{ptk_zerosum} to project on a feature, usually the output of a
benchmark model.
}
