% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{ptk_ranger}
\alias{ptk_ranger}
\title{Wrap \code{\link[ranger:ranger]{ranger::ranger()}} into a patroklos-compliant fitter}
\usage{
ptk_ranger(x, y, ...)
}
\arguments{
\item{x}{Predictor data (independent variables), alternative interface to data with formula or dependent.variable.name.}

\item{y}{Response vector (dependent variable), alternative interface to data with formula or dependent.variable.name. For survival use a \code{Surv()} object or a matrix with time and status.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A \code{ptk_ranger} S3 object, a \code{ranger} S3 object with the \code{predictions}
attribute renamed to \code{oob_predict}.
}
\description{
This function is a patroklos-compliant fitter with validated
predictions, it has a return value with a \code{oob_predict} attribute.
}
