% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_prefit}
\alias{qc_prefit}
\title{Quality control before handing over the predictor and response
matrix over to the model fitting function}
\usage{
qc_prefit(x, y)
}
\arguments{
\item{x}{A numeric matrix with dimnames holding predictors (as rows).}

\item{y}{A numeric matrix with dimnames holding responses (as rows).}
}
\description{
Check types of predictor and response matrix, their
consistency and availability.
}
\details{
Intended to be used after \code{generate_predictor()} and
\code{generate_response()}.
}
