% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_prepare}
\alias{qc_prepare}
\title{Quality control at the end of preparation step}
\usage{
qc_prepare(x, y)
}
\arguments{
\item{x}{A numeric matrix with dimnames holding predictors (as rows).}

\item{y}{A numeric matrix with dimnames holding responses (as rows).}
}
\description{
Check types of predictor and response matrix, their
consistency and availability.
}
\details{
Intended to be used after \code{generate_predictor()} and
\code{generate_response()}. \code{prepare()} and, in particular, \code{prepare_and_fit()}
call this function.
}
