% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read expression and pheno data from csv files into consistent format}
\usage{
read(directory, expr_fname, pheno_fname, patient_id_col, gene_id_col)
}
\arguments{
\item{directory}{string. The directory where both expression and pheno csv
files lie.}

\item{expr_fname}{string. The name of the expression .csv file inside \code{directory}.
See details for the expected format.}

\item{pheno_fname}{string. The name of the pheno data .csv inside \code{directory}.
See details for the expected format.}

\item{patient_id_col}{string. The name of the column in the pheno data file
that holds the patient identifiers.}

\item{gene_id_col}{string. The name of the column in the expression data file
that holds the gene identifiers.}
}
\value{
A list with a numeric matrix, named \code{expr}, and a tibble named \code{pheno}.
\code{expr} holds the expression data, with patient ids as row names and gene ids as
column names. I.e., we transpose the expression data. \code{pheno} holds the pheno data,
with the patient ids in the first column \code{patient_id_col}.
}
\description{
Read expression data into a matrix and pheno data into a tibble.
Both will hold patients as rows.
}
\details{
The pheno csv file holds the samples as rows (with \emph{unique} sample ids in a
column called \verb{patient_id_col``), the variables as columns. The expr csv file holds  the genes as rows (with *unique* gene ids in a column called }gene_id_col`), the
samples as columns.
}
