% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypes.R
\name{val_error_fun_prototype}
\alias{val_error_fun_prototype}
\title{Function interface of a validation error function}
\usage{
val_error_fun_prototype(y, y_hat)
}
\arguments{
\item{y}{numeric vector. True outcomes.}

\item{y_hat}{numeric vector. Predicted outcomes. Must have the same length as \code{y}. Samples in
\code{y} and \code{y_hat} must coincide by order.}
}
\value{
A numeric scalar.
}
\description{
This function is the prototype of a function to calculate the error of validated
predictions. It fulfills the minimal requirements in its parameters and return value to work as
the \code{val_error_fun} attribute of a \code{Model} object.
}
