% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{write_data_info}
\alias{write_data_info}
\title{Write data info to a JSON file}
\usage{
write_data_info(filename, pheno_tbl, expr_tbl, data)
}
\arguments{
\item{filename}{string. The path to the JSON file.}

\item{pheno_tbl}{tibble. The pheno data. Its format needs to comply with \code{data}
below.}

\item{expr_tbl}{tibble. The expression data. Its format needs to comply with \code{data}
below.}

\item{data}{DataSpec S3 object. Specifications on both \code{pheno_tbl} and \code{expr_tbl}.}
}
\value{
The JSON string that was written to the file.
}
\description{
Automatically generate info about the data and write it to a JSON file.
If the file already exists, the automatically inferable info into the file. If the
file is not present yet, create a sceleton, write the the info one can automtically
infer from the data into the json. You can then fill the empty fields in by writing
the json by hand. Manullay alrady filled out info will not be overwritten.
}
