% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2d.R
\name{Ass2d}
\alias{Ass2d}
\title{An R6 class to assess a model with continuous output with one metric versus another.}
\description{
Assess how well a model can predict time to event less than a certain
threshold with two metrics by thresholding the continuous output in every possible way and
then plot one metric for a binary classifier against another metric for a binary classifier.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x_metric, y_metric}}{Metrics shown on both axes.}

\item{\code{ci_level}}{For any confidence intervals calculated use this level.}

\item{\code{x_lab, y_lab}}{Axis labels.}

\item{\code{xlim, ylim}}{Axis limits.}

\item{\code{scale_x, scale_y}}{The scale of the axes.}

\item{\code{alpha}}{The alpha value for the points in the plot.}

\item{\code{colors}}{Colors used by ggplot2.}

\item{\code{theme}}{Theme applied to the plot.}

\item{\code{width, height, units}}{The size of the plot in units.}

\item{\code{dpi}}{Plot resolution in dots per inch.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{x_metric, y_metric}}{Metrics shown on both axes.}

\item{\code{x_lab, y_lab}}{Axis labels.}

\item{\code{xlim, ylim}}{Axis limits.}

\item{\code{scale_x, scale_y}}{The scale of the axes.}

\item{\code{width, height, units}}{The size of the plot in units.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ass2d-new}{\code{Ass2d$new()}}
\item \href{#method-Ass2d-assess}{\code{Ass2d$assess()}}
\item \href{#method-Ass2d-assess_center}{\code{Ass2d$assess_center()}}
\item \href{#method-Ass2d-clone}{\code{Ass2d$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-new}{}}}
\subsection{Method \code{new()}}{
Create a new Ass2d instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$new(
  x_metric,
  y_metric,
  ci_level = 0.95,
  x_lab = NULL,
  y_lab = NULL,
  xlim = c(-Inf, Inf),
  ylim = c(-Inf, Inf),
  scale_x = "identity",
  scale_y = "identity",
  alpha = 1,
  colors = NULL,
  theme = ggplot2::theme_minimal(),
  width = 7,
  height = 4,
  units = "in",
  dpi = 300
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_metric, y_metric}}{string. Metric shown on the x and y axis.
Regarding the choices:
\itemize{
\item Every combination of (\code{measure}, \code{x.measure}) one can pass to
\code{\link[ROCR:performance]{ROCR::performance()}}.
\item If \code{y_metric} is \code{"logrank"} or \code{"precision_ci"} (the upper limit of
the \code{ci_level} confidence interval of the precision, see below),
\code{x_metric} must be \code{"prevalence"} or \code{"rpp"} (rate of positive
predictions).
\item For (\code{x_metric}, \code{y_metric}) = (\code{"rank"}, "\verb{risk score}"), plot the
rank of every risk score agaisnt the risk score and color by true
risk.
}}

\item{\code{ci_level}}{numeric in [0, 1]. The level used to calculate confidence intervals.}

\item{\code{x_lab, y_lab}}{string. Axis labels. Default is \code{x_metric} and \code{y_metric}, respectively.}

\item{\code{xlim, ylim}}{numeric vector of length 2. The limits for both axes. Default is
no contraints.}

\item{\code{scale_x, scale_y}}{string or transformation object (see \code{\link[scales:new_transform]{scales::trans_new}} for the
latter). The scale of the axes, we will pass them to the \code{trans} paramter of
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}, \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}, respectively.}

\item{\code{alpha}}{numeric in [0, 1]. The alpha value for the points and lines in the
plot.}

\item{\code{colors}}{character vector. The colors to be used for the different models.
Default is \code{NULL}, which means that the default colors of \code{ggplot2} will be used.}

\item{\code{theme}}{S3 object inheriting from \code{"theme"} and \code{"gg"} (typically the return value of
\code{\link[ggplot2:theme]{ggplot2::theme()}} or a complete ggplot2 theme like \code{\link[ggplot2:ggtheme]{ggplot2::theme_light()}}).
The theme of the plot. Default is \code{ggplot2::theme_minimal()}.}

\item{\code{width}}{numeric. The width of the plot in \code{units}.}

\item{\code{height}}{numeric. The height of the plot in \code{units}.}

\item{\code{units}}{string. The units of \code{width} and \code{height}. Default is \code{"in"} (inches).}

\item{\code{dpi}}{numeric. Plot resolution in dots per inch.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Ass2d object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-assess"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-assess}{}}}
\subsection{Method \code{assess()}}{
Assess a \emph{single} model on a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$assess(
  data,
  model,
  return_type = "ggplot",
  file = NULL,
  fellow_csv = FALSE,
  quiet = FALSE,
  msg_prefix = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data. The \code{cohort} attribute
must be set.}

\item{\code{model}}{Model object. Assess this model.}

\item{\code{return_type}}{string. Either "ggplot" or "tibble". See return section for details.}

\item{\code{file}}{string. If not \code{NULL} and \verb{return_type == "ggplot}, store the resulting plot
in this file.}

\item{\code{fellow_csv}}{logical. If \code{TRUE}, \code{file} is not \code{NULL} and \code{return_type == "ggplot"},
store the plotted data in a csv named \code{file} with replaced file extension.}

\item{\code{quiet}}{logical. Whether to suppress messages.}

\item{\code{msg_prefix}}{string. Prefix for messages. Default is \code{""}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot object if \code{return_type == "ggplot"} or the tibble underlying the plot if
\code{return_type == "tibble"}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-assess_center"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-assess_center}{}}}
\subsection{Method \code{assess_center()}}{
Assess \emph{multiple} models on a data set. A wrapper around
\code{assess()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$assess_center(
  data,
  model_list,
  file = NULL,
  fellow_csv = FALSE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data. The \code{cohort} attribute must be set.}

\item{\code{model_list}}{list of Model objects. Assess these models.
We infer the \code{AssSpec2d} for the single plots in a reasonable way from it.
We skip models with the \code{continuous_output} attribute being FALSE (a plot
is an overkill in this case).}

\item{\code{file}}{string or NULL. If not \code{NULL}, store the resulting plot in this file.}

\item{\code{fellow_csv}}{logical. Whether to also store the plotted data in a csv.}

\item{\code{quiet}}{logical. Whether to suppress messages. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
