% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2d.R
\name{Ass2d}
\alias{Ass2d}
\title{An R6 class to assess a model with one metric versus another.}
\description{
Assess how well a model can predict time to event less than a certain
threshold with two metrics (like the ROC-AUC and the Brier score) to guide thresholding
the continuous to binary model output.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x_metric, y_metric}}{Metrics shown on both axes.}

\item{\code{pivot_time_cutoff}}{Assess for predicting time to event less than
pivot_time_cutoff}

\item{\code{benchmark}}{Incorporate this benchmark (from pheno data) into
assessment.}

\item{\code{file}}{Store results in this file.}

\item{\code{lambda}}{Assess model with this regularization parameter.}

\item{\code{ci_level}}{For any confidence intervals calculated use this level.}

\item{\code{fellow_csv}}{Whether to also store the plotted data in a csv
including cutoffs.}

\item{\code{scores_plot}}{Whether to additionally plot rank versus model output score.}

\item{\code{show_plots}}{Whether to print all generated plots.}

\item{\code{title}}{Plot title.}

\item{\code{x_lab, y_lab}}{Axis labels.}

\item{\code{xlim, ylim}}{Axis limits.}

\item{\code{smooth_method}}{Method to smooth the plot.}

\item{\code{smooth_benchmark}}{Whether to also smooth the benchmark.}

\item{\code{smooth_se}}{Whether to add standard error bands to the smoothed lines.}

\item{\code{scale_x, scale_y}}{The scale of the axes.}

\item{\code{vline, hline}}{Vertical/horizontal lines to be added to the plot.}

\item{\code{text_size}}{The size of in-plot text.}

\item{\code{text}}{Text label added to the plot.}

\item{\code{alpha}}{The alpha value for the points in the plot.}

\item{\code{colors}}{Colors used by ggplot2.}

\item{\code{theme}}{Theme applied to the plot.}

\item{\code{width, height, units}}{The size of the plot in units.}

\item{\code{dpi}}{Plot resolution in dots per inch.}

\item{\code{data}}{The data underlying the plots.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{x_metric, y_metric}}{Metrics shown on both axes.}

\item{\code{x_lab, y_lab}}{Axis labels.}

\item{\code{xlim, ylim}}{Axis limits.}

\item{\code{scale_x, scale_y}}{The scale of the axes.}

\item{\code{vline, hline}}{Vertical/horizontal lines to be added to the plot.}

\item{\code{width, height, units}}{The size of the plot in units.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ass2d-new}{\code{Ass2d$new()}}
\item \href{#method-Ass2d-assess}{\code{Ass2d$assess()}}
\item \href{#method-Ass2d-plot_risk_scores}{\code{Ass2d$plot_risk_scores()}}
\item \href{#method-Ass2d-assess_center}{\code{Ass2d$assess_center()}}
\item \href{#method-Ass2d-clone}{\code{Ass2d$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-new}{}}}
\subsection{Method \code{new()}}{
Create a new Ass2d instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$new(
  x_metric,
  y_metric,
  pivot_time_cutoff,
  lambda = "lambda.min",
  benchmark = NULL,
  file = NULL,
  ci_level = 0.95,
  fellow_csv = FALSE,
  scores_plot = FALSE,
  show_plots = FALSE,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  xlim = c(-Inf, Inf),
  ylim = c(-Inf, Inf),
  smooth_method = NULL,
  smooth_benchmark = FALSE,
  smooth_se = FALSE,
  scale_x = "identity",
  scale_y = "identity",
  vline = NULL,
  hline = NULL,
  text_size = 2.5,
  text = NULL,
  alpha = 0.5,
  width = 7,
  height = 4,
  colors = NULL,
  theme = NULL,
  units = "in",
  dpi = 300
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_metric}}{string. The name of the performance measure to be plotted on the x-axis.
All measures that can be passed to the \code{x.measure} parameter of \code{\link[ROCR:performance]{ROCR::performance()}} are
valid.}

\item{\code{y_metric}}{string. The name of the performance measure to be plotted on the y-axis.
\itemize{
\item All measures that can be passed to the \code{measure} parameter of \code{\link[ROCR:performance]{ROCR::performance()}} are
valid.
\item \code{"logrank"} for the p-values of a logrank test. In this case, \code{x_metric} must be
\code{"prevalence"} or \code{"rpp"}.
\item \code{"precision_ci"} for the lower bound of the \code{ci_level} (see below) confidence interval
of the precision according to Clopper-Pearson.
}}

\item{\code{pivot_time_cutoff}}{numeric. Time-to-event threshold that divides samples into a
high/low-risk (time to event below/above \code{pivot_time_cutoff}) group. Model performance
will be measured in terms of how well it can separate these two groups as a binary
classifier.}

\item{\code{lambda}}{numeric or string. The regularization parameter of the model to
evaluate.}

\item{\code{benchmark}}{string. Incorporate this benchmark (from pheno data) into the
assessment if \code{benchmark} is not \code{NULL}.}

\item{\code{file}}{string. If not \code{NULL}, store the resulting plot in this file.}

\item{\code{ci_level}}{numeric in [0, 1]. The level used to calculate confidence intervals.}

\item{\code{fellow_csv}}{logical. Whether to also store the plotted data in a csv including
cutoffs.}

\item{\code{scores_plot}}{logical. Display the ordered scores output by the model in a scatter
plot.}

\item{\code{show_plots}}{logical. Whether to show the plots after creation in an interactive
graphics device.}

\item{\code{title}}{string. The title of the plot. Default is \code{NULL}, i.e. no title.}

\item{\code{x_lab, y_lab}}{string. Axis labels. Default is \code{x_metric} and \code{y_metric}, respectively.}

\item{\code{xlim, ylim}}{numeric vector of length 2. The limits for both axes. Default is
no contraints.}

\item{\code{smooth_method}}{string or function. Smooth method to plot an additional smoothed graph.
If \code{NULL} (the default), no smoothing. Else we pass \code{smooth_method} as the \code{method}
parameter to \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}.}

\item{\code{smooth_benchmark}}{logical. Whether to also smooth the benchmark data.}

\item{\code{smooth_se}}{logical. Whether to add standard error bands to the smoothed lines.}

\item{\code{scale_x, scale_y}}{string or transformation object (see \code{\link[scales:new_transform]{scales::trans_new}} for the
latter). The scale of the axes, we will pass them to the \code{trans} paramter of
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}, \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}, respectively.}

\item{\code{vline, hline}}{list or NULL. Vertical/horizontal lines to be added to the plot. A list
holding the arguments to pass to \code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}} and \code{\link[ggplot2:geom_abline]{ggplot2::geom_hline()}},
respectively. Default is \code{NULL}, no line.}

\item{\code{text_size}}{numeric. The size of text passed to the \code{size} parameter of
\code{ggplot2::geom_text()}. The reason this attribute exists is you might want to have text
size different the global text size of the plot. See \code{vignette("ggplot2-specs", 
package = "ggplot2")} for more details.}

\item{\code{text}}{list or NULL. Text label added to the plot. A list holding the arguments to
pass to \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}. Default is \code{NULL}, no text at all.}

\item{\code{alpha}}{numeric in [0, 1]. The alpha value for the points and lines in the
plot.}

\item{\code{width}}{numeric. The width of the plot in \code{units}.}

\item{\code{height}}{numeric. The height of the plot in \code{units}.}

\item{\code{colors}}{character vector. The colors to be used for the different models.
Default is \code{NULL}, which means that the default colors of \code{ggplot2} will be used.}

\item{\code{theme}}{S3 object inheriting from \code{"theme"} and \code{"gg"} (typically the return value of
\code{\link[ggplot2:theme]{ggplot2::theme()}} or a complete ggplot2 theme like \code{\link[ggplot2:ggtheme]{ggplot2::theme_light()}}).
The theme of the plot. Default is \code{NULL}, which means the default theme of ggplot2.}

\item{\code{units}}{string. The units of \code{width} and \code{height}. Default is \code{"in"} (inches).}

\item{\code{dpi}}{numeric. Plot resolution in dots per inch.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Ass2d object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-assess"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-assess}{}}}
\subsection{Method \code{assess()}}{
Assess a \emph{single} model (with multiple splits) on a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$assess(data, model, quiet = FALSE, msg_prefix = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data. Data must already be read in and
\code{cohort} attribute set.}

\item{\code{model}}{ModelSpec object. Assess this model, multiple splits are supported.}

\item{\code{quiet}}{logical. Whether to suppress messages.}

\item{\code{msg_prefix}}{string. Prefix for messages. Default is \code{""}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
We add the data underlying the plots to Ass2d object as a new attribute
named \code{data}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-plot_risk_scores"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-plot_risk_scores}{}}}
\subsection{Method \code{plot_risk_scores()}}{
Plot rank versus model score on a data set. You can do this
on the fly, i.e. you actually initialize the Ass2d object for another
purpose and this method will infer reasonable defaults for the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$plot_risk_scores(data, model, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Plot the scores for this data.}

\item{\code{model}}{Model object. Plot the scores for this model.}

\item{\code{quiet}}{logical. Whether to suppress messages. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-assess_center"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-assess_center}{}}}
\subsection{Method \code{assess_center()}}{
Assess \emph{multiple} models (with multiple splits) on a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$assess_center(
  data,
  model_list,
  model_tree_mirror = c("models", "results"),
  comparison_plot = TRUE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{DataSpec object. Assess on this data set.}

\item{\code{model_list}}{list of ModelSpec objects. Assess these models.
We infer the \code{AssSpec2d} for the single plots in a reasonable way from it.}

\item{\code{model_tree_mirror}}{character vector of length 2. Store the single plots in
the same directory as the corresponding model for the train cohort and the
comparison plot at \code{file}. When doing the analogous for the test data, replace
the first element of \code{model_tree_mirror} by the second element in every file path.}

\item{\code{comparison_plot}}{logical. Whether to generate a plot holding the assessment
of all models. Default is \code{TRUE}.}

\item{\code{quiet}}{logical. Whether to suppress messages. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ass2d-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ass2d-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ass2d$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
