% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass2d_helpers.R
\name{assess_2d}
\alias{assess_2d}
\title{Assess a model on a single data set in terms how well it filters high-risk
samples in a 2d plot}
\usage{
assess_2d(
  expr_mat,
  pheno_tbl,
  data,
  model,
  ass2d,
  quiet = FALSE,
  msg_prefix = ""
)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix. Rows are samples, columns
are genes, their identifiers are given as row and column names, respectively.}

\item{pheno_tbl}{tibble. The phenotype table. Observations are samples, columns
are variables.}

\item{data}{DataSpec object specifying \code{expr_mat} and \code{pheno_tbl}. See the
constructor \code{DataSpec()} for details.}

\item{model}{ModelSpec object. The model to be assessed. See the constructor
\code{ModelSpec()} for details.}

\item{ass2d}{AssSpec2d object. The specifications for the plot. See
the constructor \code{AssSpec2d()} for details. The \code{pivot_time_cutoff} attribute of
\code{ass2d} will override the same in \code{model} if both are given.}

\item{quiet}{logical. Whether to suppress messages. Default is \code{FALSE}.}

\item{msg_prefix}{string. Prefix for messages. Default is \code{""}.}
}
\value{
A AssSpec2d object. \code{ass2d} with one additional attribute named
\code{data} holding the data underlying the plots.
}
\description{
Given a data set and a model, assess how well the model can filter
high-risk patients. This includes:
\enumerate{
\item A 2D scatter plot of two performance metrics, namely the \code{x_metric} and \code{y_metric}
attributes of \code{ass2d}.
\item A plot of the risk scores of the model.
Supports multiple splits into train and test cohorts, but not multiple time cutoffs
or multiple models. For the this, see \code{\link[=assess_2d_center]{assess_2d_center()}}.
}
}
\details{
The AssSpec2d class is tailored for this function, so see its constructor
\code{\link[=AssSpec2d]{AssSpec2d()}} for details.
}
