% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_model.R
\name{assess_model}
\alias{assess_model}
\title{Assess a model on a single data set in terms how well it filters high-risk
samples}
\usage{
assess_model(
  expr_mat,
  pheno_tbl,
  data_spec,
  model_spec,
  perf_plot_spec,
  quiet = FALSE,
  msg_prefix = ""
)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix. Rows are samples, columns
are genes, their identifiers are given as row and column names, respectively.}

\item{pheno_tbl}{tibble. The phenotype table. Observations are samples, columns
are variables.}

\item{data_spec}{DataSpec object specifying \code{expr_mat} and \code{pheno_tbl}. See the
constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec object. The model to be assessed. See the constructor
\code{ModelSpec()} for details.}

\item{perf_plot_spec}{PerfPlotSpec object. The specifications for the plot. See
the constructor \code{PerfPlotSpec()} for details. The \code{pivot_time_cutoff} attribute of
\code{perf_plot_spec} will override the same in \code{model_spec} if both are given.}

\item{quiet}{logical. Whether to suppress messages. Default is \code{FALSE}.}

\item{msg_prefix}{string. Prefix for messages. Default is \code{""}.}
}
\value{
A PerfPlotSpec object. \code{perf_plot_spec} with one additional attribute named
\code{data} holding the data underlying the plots.
}
\description{
Given a data set and a model, assess how well the model can filter
high-risk patients. This includes:
\enumerate{
\item A 2D scatter plot of two performance metrics, namely the \code{x_metric} and \code{y_metric}
attributes of \code{perf_plot_spec}.
\item A plot of the risk scores of the model.
Supports multiple splits into train and test cohorts, but not multiple time cutoffs
or multiple models. For the this, see \code{\link[=assessment_center]{assessment_center()}}.
}
}
\details{
The PerfPlotSpec class is tailored for this function, so see its constructor
\code{\link[=PerfPlotSpec]{PerfPlotSpec()}} for details.
}
