% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass_scalar.R
\name{AssScalar}
\alias{AssScalar}
\title{An R6 class to assess a model with a scalar metric}
\description{
Assess how well a model can predict time to event less than a certain
threshold with a \emph{scalar} metric (like the ROC-AUC).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metric}}{Scalar metric.}

\item{\code{pivot_time_cutoff}}{Assess classifying time to event less than
pivot_time_cutoff.}

\item{\code{lambda}}{Assess the model with regularization parameter lambda.}

\item{\code{benchmark}}{Incorporate the benchmark into the assessment.}

\item{\code{round_digits}}{Round the results in tables to round_digits digits
after the point.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AssScalar-new}{\code{AssScalar$new()}}
\item \href{#method-AssScalar-assess}{\code{AssScalar$assess()}}
\item \href{#method-AssScalar-assess_center}{\code{AssScalar$assess_center()}}
\item \href{#method-AssScalar-clone}{\code{AssScalar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-new"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-new}{}}}
\subsection{Method \code{new()}}{
Create a new AssScalar instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$new(
  metric,
  pivot_time_cutoff,
  lambda = "lambda.min",
  benchmark = NULL,
  file = NULL,
  round_digits = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric}}{character. Name of the function used to calculate the metric. It must
take two numeric vectors of the same length as arguments:
\enumerate{
\item \code{predicted}, the predictions by a model, higher values indicate a more positive
prediction,
\item \code{actual}, the true values with the positive class coded as 1, the negative one as 0.
}}

\item{\code{pivot_time_cutoff}}{numeric. Assess for classifying time to event <
\code{pivot_time_cutoff}.}

\item{\code{lambda}}{Assess the model belonging to this \code{lambda} in the cross validation
(if there was one).}

\item{\code{benchmark}}{character or NULL. If not NULL, include \code{benchmark} (the name of column
in the pheno data) in the assessment.}

\item{\code{file}}{string or NULL. The name of the csv file to save the results to. If \code{NULL},
the results are not saved. If you supply an AssScalar object to \code{AssScalar_assess_center()},
specify the path for the \emph{train} cohort.}

\item{\code{round_digits}}{numeric. The number of digits to round the results to. Default is \code{3}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AssScalar object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-assess"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-assess}{}}}
\subsection{Method \code{assess()}}{
Assess a \emph{single} model (with multiple splits) on a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$assess(data, model, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data. Data must already be read in
and \code{cohort} attribute set.}

\item{\code{model}}{Model object. Assess this model, multiple splits are supported.}

\item{\code{quiet}}{logical. Whether to suppress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The AssScalar S3 class is tailored for this function.
}

\subsection{Returns}{
numeric vector. For every split index the desired metric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-assess_center"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-assess_center}{}}}
\subsection{Method \code{assess_center()}}{
Assess \emph{multiple} models (with multiple splits) on a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$assess_center(
  data,
  model_list,
  model_tree_mirror = c("models", "results"),
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data.
If \code{data$cohort} is \code{NULL}, assess on the test cohort.}

\item{\code{model_list}}{list of Model objects. Assess these models.}

\item{\code{model_tree_mirror}}{character vector of length 2. If you want to store the
resulting tibbles, get the test-cohort tibble's file name by mirroring
\code{ass_scalar$file} according to \code{model_tree_mirror} (see \code{\link[=mirror_path]{mirror_path()}}).}

\item{\code{quiet}}{logical. Whether to suppress messages.}

\item{\code{ass_scalar}}{AssScalar S3 object. See the constructor \code{\link[=AssScalar]{AssScalar()}} for more
details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble. Every row holds the metric (and more analysis across the splits
like standard deviation, minimum, maximum value) for one model in \code{model_list}
and every time cutoff specified for this model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
