% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ass_scalar.R
\name{AssScalar}
\alias{AssScalar}
\title{An R6 class to assess a model with a scalar metric}
\description{
Assess how well a model can predict time to event less than a certain
threshold with a \emph{scalar} metric (like the ROC-AUC).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metrics}}{Assess the model for these scalar metrics. Check out the
initializer for possible choices.}

\item{\code{prev_range}}{For metrics that need thresholding only consider
thresholds that yield a prevalence in this range.}

\item{\code{confidence_level}}{Confidence level alpha.}

\item{\code{benchmark}}{Incorporate the benchmark into the assessment.}

\item{\code{round_digits}}{Round the results in tables to round_digits digits
after the point.}

\item{\code{file}}{Save the resulting tibble to this csv file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AssScalar-new}{\code{AssScalar$new()}}
\item \href{#method-AssScalar-assess}{\code{AssScalar$assess()}}
\item \href{#method-AssScalar-assess_center}{\code{AssScalar$assess_center()}}
\item \href{#method-AssScalar-clone}{\code{AssScalar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-new"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-new}{}}}
\subsection{Method \code{new()}}{
Create a new AssScalar instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$new(
  metrics = c("auc", "accuracy", "precision", "prevalence", "precision_ci_ll", "hr",
    "hr_ci_ll", "hr_ci_ul", "hr_p", "n_true", "perc_true", "n_samples", "logrank",
    "threshold"),
  prev_range = c(0, 1),
  confidence_level = 0.95,
  benchmark = NULL,
  file = NULL,
  round_digits = 3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metrics}}{character. Assess the model for these metrics. For
currently offered choices see "Usage" above. If you have a model with
non-binary output (like a logistic regression), we choose a threshold
by maximizing the left-most metric in \code{metrics} \emph{that is made for
classifiers with binary output} (e.g. precision within prev_range below).
If this cannot be done reasonably, throw an error. Make sure that \code{hr} precedes
\code{hr_ci_ll}, \code{hr_ci_ul} and \code{hr_p} in \code{metrics}.}

\item{\code{prev_range}}{numeric numeric vector of length 2. For metrics that
need thresholding only consider thresholds that yield a prevalence in
this range.}

\item{\code{confidence_level}}{numeric. The confidence level alpha (e.g. for confidence
intervals).}

\item{\code{benchmark}}{character or NULL. If not NULL, include \code{benchmark}
(the name of column in the pheno data) in the assessment. Not
implemented yet.}

\item{\code{file}}{string or NULL. The name of the csv file to save the
results to for the \emph{train} cohort.
the results are not saved.}

\item{\code{round_digits}}{numeric. The number of digits to round the results to. Default is \code{3}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AssScalar object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-assess"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-assess}{}}}
\subsection{Method \code{assess()}}{
Assess a \emph{single} model on a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$assess(data, model, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data. Data must already be read in
and \code{cohort} attribute set.}

\item{\code{model}}{Model object. Assess this model.}

\item{\code{quiet}}{logical. Whether to suppress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The AssScalar S3 class is tailored for this function.
}

\subsection{Returns}{
numeric matrix. The rows correspond to the \code{model} and optionally
the benchmark, the columns correspond to \code{self$metrics}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-assess_center"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-assess_center}{}}}
\subsection{Method \code{assess_center()}}{
Assess \emph{multiple} models on a data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$assess_center(
  data,
  model_list,
  mirror = c("models", "results"),
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data object. Assess on this data.
If \code{data$cohort} is \code{NULL}, assess on the test cohort.}

\item{\code{model_list}}{list of Model objects. Assess these models.}

\item{\code{mirror}}{character vector of length 2. If you want to store the
resulting tibbles, get the test-cohort tibble's file name by mirroring
\code{ass_scalar$file} according to \code{model_tree_mirror}, i.e. replacing
\code{model_tree_mirror[1]} by \code{model_tree_mirror[2]} in the \code{file}
attribute.}

\item{\code{quiet}}{logical. Whether to suppress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble. Every row stands for one model.
\itemize{
\item If \code{length(self$metrics) == 1}, the columns hold statistics on the
single metric (like mean, standard deviation).
\item Otherwise columns correspond to \code{self$metrics}.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssScalar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AssScalar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssScalar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
