% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Data}
\alias{Data}
\title{An R6 class for the data}
\description{
A Data object holds the phenotype and expression data belonging to
a data set. It specifies the names of important features in the columns, reads from
and writes to csv files and takes care of splitting the data into train and test
cohort.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{How to refer to the data in plots and messages.}

\item{\code{directory}}{Directory where the expression and pheno csv files lie.}

\item{\code{train_prop}}{Proportion of samples in the training cohort.}

\item{\code{expr_mat}}{The expression matrix.}

\item{\code{pheno_tbl}}{The phenotype tibble.}

\item{\code{expr_file}}{Name of the expression csv file inside \code{directory}.}

\item{\code{pheno_file}}{Name of the pheno data csv inside \code{directory}.}

\item{\code{cohort}}{The cohort of the data, train or test.}

\item{\code{cohort_col}}{Find the cohort of a sample in this column of the
pheno data.}

\item{\code{patient_id_col}}{The name of the column in the pheno data that holds
the patient identifiers.}

\item{\code{time_to_event_col}}{The name of the column in the pheno data that holds the
time-to-event values.}

\item{\code{event_col}}{The name of the column in the pheno data that holds
the event status encoded as 1 = occurrence, 0 = censoring.}

\item{\code{benchmark_col}}{The name of the column in the pheno data that holds the
benchmark risk score (like the IPI).}

\item{\code{gene_id_col}}{The name of the column in the expression data that holds
the gene identifiers.}

\item{\code{split_col_prefix}}{Column-name prefix for those columns holding splits into
train and test cohort.}

\item{\code{pivot_time_cutoff}}{Pivotal time cutoff for the analysis and, explicitly,
for splitting the data into the train and test cohort.}

\item{\code{imputer}}{Function handling NAs in the predictor matrix.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Data-new}{\code{Data$new()}}
\item \href{#method-Data-read}{\code{Data$read()}}
\item \href{#method-Data-prepare}{\code{Data$prepare()}}
\item \href{#method-Data-survival_quantiles}{\code{Data$survival_quantiles()}}
\item \href{#method-Data-qc_preprocess}{\code{Data$qc_preprocess()}}
\item \href{#method-Data-clone}{\code{Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-new"></a>}}
\if{latex}{\out{\hypertarget{method-Data-new}{}}}
\subsection{Method \code{new()}}{
A Data object specifies the location and format of the expression
and pheno data of a single data set. This enables reading and preparing the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$new(
  name,
  directory,
  train_prop,
  pivot_time_cutoff,
  time_to_event_col,
  cohort,
  event_col = "progression",
  cohort_col = "split_1",
  benchmark_col = NULL,
  expr_file = "expr.csv",
  pheno_file = "pheno.csv",
  patient_id_col = "patient_id",
  gene_id_col = "gene_id",
  split_col_prefix = "split_",
  imputer = mean_impute
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string. A telling name for the data set (e.g. \code{"Schmitz et al. 2018"}).}

\item{\code{directory}}{string. The directory where both expression and pheno csv
files lie.}

\item{\code{train_prop}}{numeric in (0, 1). The proportion of samples to draw for the
training cohort when a new split into train and test cohort is performed.}

\item{\code{pivot_time_cutoff}}{numeric. Pivotal time cutoff for the analysis and, explicitly,
for splitting the data into the train and test cohort: If a numeric in (0, 1), preserve
the proportion of individuals below and above \code{time_cutoff} in both cohorts. Default
is \code{NULL}, i.e. no further constraints on splitting.}

\item{\code{time_to_event_col}}{string. The name of the column in the pheno data that holds the
time-to-event values.}

\item{\code{cohort}}{string. At the end of preparing the data, subset it to those
samples whose value in the split column matches \code{cohort}. Default is \code{NULL},
in which case some methods will infer a reasonable value (like \code{"train"}
for training, \code{"test"} for assessment), some will throw an error.}

\item{\code{event_col}}{string. The name of the column in the pheno data that holds
the event status encoded as 1 = occurrence, 0 = censoring.}

\item{\code{cohort_col}}{string. The name of the column in the pheno data that holds
the cohort a sample belongs to.}

\item{\code{benchmark_col}}{string or \code{NULL}. The name of the column in the pheno data that
holds the benchmark risk score (like the IPI).}

\item{\code{expr_file}}{string. The name of the expression csv file inside \code{directory}.
Default is \code{"expr.csv"}. See details for the expected format.}

\item{\code{pheno_file}}{string. The name of the pheno data csv inside \code{directory}.
Default is \code{"pheno.csv"}. See details for the expected format.}

\item{\code{patient_id_col}}{string. The name of the column in the pheno data that holds
the patient identifiers.}

\item{\code{gene_id_col}}{string. The name of the column in the expression data that holds
the gene identifiers.}

\item{\code{split_col_prefix}}{string. Column-name prefix for those columns in the pheno
tibble holding splits into cohorts (like train and test). Some of
these columns may be present in the pheno data already, non-present ones will be
added during the training process if required. Default is \code{"split_"}, i.e.,
split columns are named \code{"split_1"}, \code{"split_2"}, etc.}

\item{\code{imputer}}{function or NULL. Function handling NAs in the predictor matrix.
It takes a single argument \code{x}, the predictor matrix, a numeric with categorical
variables dichotomized to binary dummy variables, and returns a matrix of the
same shape with non-NAs left untouched. Default is \code{\link[=mean_impute]{mean_impute()}}.
NULL means no imputation (or applying the identity function on \code{x}).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The pheno csv file holds the samples as rows (with \emph{unique} sample ids in the
first (character) column called \code{patient_id_col}), the variables as columns.

The expression csv file holds the genes as rows (with \emph{unique} gene ids in the first
(character) column called \code{gene_id_col}), the samples as columns.

While the computational representation of both expression and pheno data will
change over the course of the pipeline, a Data object will hold timeless
information on the data.
}

\subsection{Returns}{
A Data object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-read"></a>}}
\if{latex}{\out{\hypertarget{method-Data-read}{}}}
\subsection{Method \code{read()}}{
Read expression data into the matrix \code{expr_mat} and pheno
data into the tibble \code{pheno_tbl}. Both will hold patients as rows.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-prepare"></a>}}
\if{latex}{\out{\hypertarget{method-Data-prepare}{}}}
\subsection{Method \code{prepare()}}{
Prepare the read-in data for a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$prepare(model, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A Model object.}

\item{\code{quiet}}{logical. If TRUE, suppress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
You need to set the cohort before calling this method.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-survival_quantiles"></a>}}
\if{latex}{\out{\hypertarget{method-Data-survival_quantiles}{}}}
\subsection{Method \code{survival_quantiles()}}{
Calculate the quantiles of the survival times.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$survival_quantiles(round_digits = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{round_digits}}{integer. Round the numbers in the returned tibble
to this number of digits after the point.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
We take censoring into account.
}

\subsection{Returns}{
A tibble with two columns. For each quantile q in the first
column, the time-to-event value in the second column.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-qc_preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-Data-qc_preprocess}{}}}
\subsection{Method \code{qc_preprocess()}}{
Quality control at the end of preprocessing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$qc_preprocess(expr_tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr_tbl}}{A tibble with the expression data, the product of
reading in an expression csv file as described in the details of
the initialize() method.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the expression and pheno tibble are consistent
with the other attributes of the Data object. You typically call this
method at the end of preprocessing, and the read() method calls it.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
