% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_spec.R
\name{DataSpec}
\alias{DataSpec}
\title{Construct a DataSpec S3 object}
\usage{
DataSpec(
  patient_id_col = "patient_id",
  pfs_col = "pfs_years",
  progression_col = "progression",
  gene_id_col = "gene_id"
)
}
\arguments{
\item{patient_id_col}{string. The name of the column in the pheno data that holds
the patient identifiers. Default is \code{"patient_id"}.}

\item{pfs_col}{string. The name of the column in the pheno data that holds the
progression-free survival (PFS) values. Default is \code{"pfs_years"}.}

\item{progression_col}{string. The name of the column in the pheno data that holds
the progression status encoded as 1 = progression, 0 = no progression. Default is
\code{"progression"}.}

\item{gene_id_col}{string. The name of the column in the expression data that holds
the gene identifiers. Default is \code{"gene_id"}.}
}
\value{
A DataSpec object.
}
\description{
A DataSpec object holds the pieces of information key to reading and
preparing the expression and pheno data, namely:
}
\details{
Note that the appearance of the pheno data changes over the course of the
pipeline: first, it is csv file, later it is a tibble.
}
