% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensures.R
\name{ensure_patients_match}
\alias{ensure_patients_match}
\title{Ensure patients in expression and pheno data match}
\usage{
ensure_patients_match(
  stage,
  expr,
  pheno,
  patient_id_col = NULL,
  gene_id_col = NULL
)
}
\arguments{
\item{stage}{string. The pipeline stage we are in. One of \code{"preprocessing"},
\code{"after_generate_xy"}.}

\item{expr}{tibble or numeric matrix.
\itemize{
\item If \code{stage == "preprocessing"}: \code{expr} is a tibble with genes as rows and
patients as columns. The column \code{gene_id_col} (see below) holds the gene
dentifiers, all other columns hold the patient identifiers.
\item If \code{stage == "after_generate_xy"}: \code{expr} is a numeric matrix with patients
as rows and genes as columns. The row names hold the patient identifiers, the
column names hold the gene identifiers.
}}

\item{pheno}{tibble or numeric matrix. The pheno data, with patients as rows
and variables as columns. The column \code{patient_id_col} (see below) holds the
patient identifiers.
\itemize{
\item If \code{stage == "preprocessing"}: \code{pheno} is a tibble.
\item If \code{stage == "after_generate_xy"}: \code{pheno} is a numeric matrix as returned
by \code{generate_response()}.
}}

\item{patient_id_col}{string. The name of the column in \code{pheno} that holds
the patient identifiers. Default is \code{NULL}.}

\item{gene_id_col}{string. The name of the column in \code{expr} that holds the
gene identifiers. \code{patient_id_col} and \code{gene_id_col} are only used if
\code{stage == "preprocessing"}. Default is \code{NULL}.}
}
\value{
A list with two elements, \code{expr} and \code{pheno}. Matching and sorted
expression and pheno data, i.e. the vectors of patient identifiers are
identical.
}
\description{
Ensure that the patients in the expression and pheno data match
via subsetting and sorting depenendent on pipeline stage.
}
