% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit and store a model}
\usage{
fit(x, y, model_spec, save_plots = TRUE)
}
\arguments{
\item{x}{numeric matrix. The predictor matrix with samples in rows and their names
as row names.}

\item{y}{numeric matrix. The response matrix with samples in rows and their names
as row names.}

\item{model_spec}{ModelSpec S3 object. Specifications on the model. See the the
constructor \code{ModelSpec()} for details.}

\item{save_plots}{logical. Whether to save the lambda-vs-deviance plot in
\code{model_spec$save_dir}. Default is \code{TRUE}.}
}
\value{
A \code{zeroSum.fit} object, with a \code{ModelSpec} object appended.
}
\description{
Given a model and data in a model-specific form, fit this model
and store it
}
