% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xy.R
\name{generate_predictor}
\alias{generate_predictor}
\title{Generate the predictor matrix in a model-specific way}
\usage{
generate_predictor(expr_mat, pheno_tbl, data_spec, model_spec)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression data, with patients as rows
and genes as columns.}

\item{pheno_tbl}{tibble. The pheno data, with patients as rows and variables as
columns.}

\item{data_spec}{DataSpec S3 object. Specifications on the data (\code{expr_mat} and
\code{model_spec}). See the the constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec S3 object. Specifications on the model to prepare for.
See the the constructor \code{ModelSpec()} for details.}
}
\value{
A numeric matrix with patients as rows and variables as columns.
}
\description{
Generate the numeric predictor matrix from the expression and
(possibly) pheno data for a certain model
}
