% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xy.R
\name{generate_predictor}
\alias{generate_predictor}
\title{Generate the predictor matrix in a model-specific way}
\usage{
generate_predictor(
  expr_mat,
  pheno_tbl,
  patient_id_col,
  include_from_continuous_pheno = NULL,
  include_from_discrete_pheno = NULL
)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression data, with patients as rows
and genes as columns.}

\item{pheno_tbl}{tibble. The pheno data, with patients as rows and variables as
columns.}

\item{include_from_continuous_pheno}{vector of strings. The names of the continuous
variables in \verb{pheno_tbl to include in the predictor matrix. The values will be coerced to numeric. Default is }NULL`, which means no continuous pheno
variables will be included.}

\item{include_from_discrete_pheno}{vector of strings. The names of the discrete
variables in \code{pheno_tbl} to be include in the predictor matrix. A discrete
variable with n levels will be converted to n-1 dummy binary variables. Default is
\code{NULL}, which means no discrete pheno variables will be included.}
}
\value{
A numeric matrix with patients as rows and variables as columns.
}
\description{
Generate the numeric predictor matrix from the expression and
(possibly) pheno data for a certain model
}
