% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xy.R
\name{generate_response}
\alias{generate_response}
\title{Generate the response matrix in a model-specific way}
\usage{
generate_response(
  pheno_tbl,
  model,
  patient_id_col = "patient_id",
  pfs_col = "pfs_years",
  progression_col = "progression",
  pfs_leq = 2
)
}
\arguments{
\item{pheno_tbl}{tibble. The pheno data, with patients as rows and variables as
columns.}

\item{model}{string. The name of the model. One of \code{c("lasso-zerosum", "cox-lasso-zerosum")}.}

\item{patient_id_col}{string. The name of the column in the pheno data file that holds
patient identifiers. Default is \code{"patient_id"}.}

\item{pfs_col}{string. The name of the column in the pheno data file that holds the
progression-free survival (PFS) data. Default is \code{"pfs_years"}.}

\item{progression_col}{string. The name of the column in the \code{pheno_tbl} that holds
the progression status encoded as 1 = progession, 0 = no progression.
Default is \code{"progression"}.}

\item{pfs_leq}{numeric. Categorize patients with progression-free survival (PFS) less than
or equal \code{pfs_leq} as high-risk. Only necessary if model discretizes response (PFS).
Default is 2.0.}
}
\value{
A numeric matrix with patients as rows and variables as columns.
}
\description{
Generate the numeric response matrix from the pheno data for a
certain model
}
