% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{ModelSpec}
\alias{ModelSpec}
\title{Construct a ModelSpec S3 object}
\usage{
ModelSpec(
  name,
  directory,
  fitter,
  split_index,
  time_cutoffs,
  optional_fitter_args = NULL,
  response_type = c("binary", "survival_censored"),
  response_colnames = c("time", "status"),
  include_from_continuous_pheno = NULL,
  include_from_discrete_pheno = NULL,
  append_to_includes = "++",
  create_directory = TRUE,
  plot_fname = "training_error.pdf",
  plot_ncols = 2,
  fit_fname = "fit_obj.rds"
)
}
\arguments{
\item{name}{string. A telling name for the model.}

\item{directory}{string. The directory to store the models in. For every value in
\code{time_cutoffs}, find the corresponding model in a subdirectory named after this value.}

\item{fitter}{function. The model fitting function to be used. Must take \code{x} and
\code{y} as first two positional arguments. Further arguments can be passed via
\code{optional_fitter_args} (below). Its return value must be an S3 object with a \code{plot()}
method, and (ideally, for assessment) with a \code{predict()} method. Default is \code{NULL}.}

\item{split_index}{integer vector. Split the given data into training and test samples
\code{length(split_index)} times, i.e., every index in \code{split_index} will get its own split.}

\item{time_cutoffs}{numeric vector.
\itemize{
\item If \code{response_type == "survival_censored"}: For every value in \code{time_cutoffs}, censor
all patients where the event ouccured after \code{time_cutoffs} at this value and train the
specified model.
\item If \code{response_type == "binary"}: For every value in \code{time_cutoffs}, binarize the
outcome depending on whether it occured before or after this value, and train the
specified model.
}}

\item{optional_fitter_args}{list. Optional arguments passed to \code{fitter}, e.g. alpha
in case of an elastic net. Default is \code{list()}, i.e., no arguments other than \code{x}, \code{y}
passed to \code{fitter}.}

\item{response_type}{string. The type of response to be used. One of \code{"binary"} or
\code{"survival_censored"}. Default is \code{NULL}.}

\item{response_colnames}{string vector of length 2. If \code{response_type == "survival_censored"},
use as column names for the response matrix.
\itemize{
\item The first element is the name of the column holding the time until the event or
censoring, and
\item the second one is the anme of the column holding the event status (1 = event, 0 =
censoring).
Default is \code{c("time", "status")}.
}}

\item{include_from_continuous_pheno}{vector of strings. The names of the
\emph{continuous} variables in the pheno data (to be) included in the predictor matrix. The
values will be coerced to numeric. Default is \code{NULL}, which means no continuous pheno
variables are or will be included.}

\item{include_from_discrete_pheno}{vector of strings. The names of the \emph{discrete}
variables in the pheno data (to be) included in the predictor matrix. A discrete
variable with n levels will be converted to n-1 binary variables. Default is \code{NULL},
which means no discrete pheno variables are or will be included.}

\item{append_to_includes}{string. Append this to the names of features from the pheno
data when adding them to the predictor matrix. Default is \code{"++"}.}

\item{create_directory}{logical. Whether to create \code{directory} if it does not exist, yet.
Default is \code{TRUE}.}

\item{plot_fname}{string. Store the plot resulting from \code{plot(fit_obj)} in \code{directory}
under this name. Default is \code{"training_error.pdf"}.}

\item{plot_ncols}{integer. The number of columns in the plot. Default is \code{2}.}

\item{fit_fname}{string. The name of the model-fits file inside \code{directory}.
Default is \code{"fit_obj.rds"}.}
}
\value{
A ModelSpec S3 object.
}
\description{
A ModelSpec object holds all the tools and information needed to fit and
store models. Most importantly, it is passed as an argument to \code{\link[=training_camp]{training_camp()}}.
Its base object is a list.
}
\details{
Strictly speaking, one \code{ModelSpec} instance holds the instructions to fit
\code{length(time_cutoffs) * length(split_index)} models. In terms of storing and assessing models,
we consider the models obtained via repeated splitting according to \code{split_index} as one
model; repeated splitting serves the purpose of getting more reliable estimates of its
performance. We view models obtained via different values of \code{time_cutoffs}, in
contrast, as different models; e.g., we can compare them against one another in an
assessment.
}
