% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cv.R
\name{nested_fit}
\alias{nested_fit}
\title{Construct a nested_fit S3 object}
\usage{
nested_fit(model1, model2, metric_grid, best_hyperparams)
}
\arguments{
\item{model1}{An S3 object with a \code{predict} method. The early model.}

\item{model2}{An S3 object with a \code{predict} method. The late model.}

\item{metric_grid}{named numeric matrix. Entry (i, j) is the validated metric
of the nested model with the early model having the i-th and the late model
having the j-th of their respective hyperparameters (which are given as
dimnames).}

\item{best_hyperparams}{A named list with the best hyperparameters.}
}
\value{
An S3 object with class `nested_fit'.
}
\description{
Construct a nested_fit object that holds the early and late model
plus their hyperparameters.
}
\details{
Nest two models. The first model takes part of the features (those
not matching pattern) to make a prediction that we in turn feed together with
the rest of the features to the second model. This gives us a final prediction.
Both models must be S3 objects with a \code{predict} method.
}
