% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cv.R
\name{nested_pseudo_cv}
\alias{nested_pseudo_cv}
\title{Nested cross-validation for second-stage OOB predictions}
\usage{
nested_pseudo_cv(
  x,
  y,
  fitter1,
  fitter2,
  hyperparams1,
  hyperparams2,
  n_folds = 10,
  oob = c(FALSE, TRUE)
)
}
\arguments{
\item{x}{A numeric matrix holding the predictor features: rows are samples and
columns are features.}

\item{y}{A numeric vector holding the response variable.}

\item{fitter1}{A function that fits the early model. It must be able to perform
a cross validation and return an S3 class with
\itemize{
\item a \code{cv_predict} attribute, a list of vectors, for every lambda the
cross-validated predictions.
\item a \code{lambda} attribute, a numeric vector holding the lambda values, and
\item a \code{predict} method.
}}

\item{fitter2}{A function that fits the late model. It must return an S3 object
with a \code{predict} method.}

\item{hyperparams1}{A named list with hyperpamaters we will pass to \code{fitter1}
to obtain cross-validated predictions from the early model.}

\item{hyperparams2}{A named list with hyperparameters for the late model.
Unlike \code{hyperparams1}, we call \code{fitter2} for every combination of values in
\code{hyperparams2} and lambda value from \code{fitter1}.}

\item{n_folds}{An integer specifying the number of folds for the cross-validation.}

\item{oob}{A logical vector of length 2. If the first element is \code{TRUE}, train
the late model on out-of-bag, else cross-validated predictions from the early
model. If the second element is \code{TRUE}, evaluate the entire model on OOB,
else cross-validated predictions.}
}
\value{
An S3 object with class \code{nested_fit}, the model with the best
performance according to the out-of-bag (OOB) predictions based on cross-validated
predictions from the early model.
}
\description{
Perform a nested cross-validation for a late-integration scheme,
i.e., perform a cross-validation for the early model and then train second-stage
models on cross-validated predictions and evaluate the entire models via
out-of-bag (OOB) predictions of the second-stage models (usually a random
forest).
}
\details{
This function does hyperparameter tuning for a nested model, i.e.,
a so-called early model makes predictions from the high-dimensional part of
data (e.g. RNA-seq, Nanostring), we then provide these predictions as a
one-dimensional feature together with new features to a late model. Both the
early and late model try to predict \code{y}. To not provide the late model overly
optimistic (since overfitted) predictions during training, we feed its
training algorithm with values comparable to those we would observe for
independent test samples, i.e. either cross-validated or out-of-bage (OOB)
predictions. To evaluate the overall model, we do a second cross-validation
or use OOB predictions.
Note that the predictions we get for the nested model are not predictions as
one would observe for independent test samples:
Let's fix sample i. We get the OOB/CV prediction for sample i from models/
a model whose training algorithm didn't see sample i itself. But it probably
saw the prediction for a sample j != i according to an early model whose
training algorithm had seen sample i. Hence the term "pseudo" in the name of
this function. This heuristic saves a factor \code{n_folds} computation time
compared to a full nested cross-validation.
}
