% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_plot_spec.R
\name{PerfPlotSpec}
\alias{PerfPlotSpec}
\title{Create a PerfPlotSpec object}
\usage{
PerfPlotSpec(
  fname,
  x_metric,
  y_metric,
  pfs_leq = NULL,
  benchmark = "ipi",
  fellow_csv = TRUE,
  scores_plot = TRUE,
  x_lab = NULL,
  y_lab = NULL,
  width = 7,
  height = 4,
  units = "in",
  title = NULL,
  colors = NULL,
  show_plots = FALSE
)
}
\arguments{
\item{fname}{string. The name of the file to save the plot to.}

\item{x_metric}{string. The name of the performance measure to be plotted on the x-axis.
All measures that can be passed to the \code{x.measure} parameter of \code{ROCR::performance()} are
valid.}

\item{y_metric}{string. The name of the performance measure to be plotted on the y-axis.
All measures that can be passed to the \code{measure} parameter of \code{ROCR::performance()} are
valid.}

\item{pfs_leq}{numeric. Progression-free survival threshold that divides samples into
high-risk (PFS < \code{pfs_leq}) and low-risk (PFS >= \code{pfs_leq}) group. Model performance
will be measured in terms of how well it can separate these two groups as a binary
classifier. Default is \code{2.0}.}

\item{benchmark}{string. Column in the test pheno holding the numeric benchmark values.
Default is \code{"ipi"} (international prognostic index for DLBCL).}

\item{fellow_csv}{logical. If passed to \code{compare_models()}, whether to also create a
csv file for every model-data pair. Default is \code{TRUE}.}

\item{scores_plot}{logical. Display the ordered scores output by the model in a scatter
plot. Default is \code{TRUE}.}

\item{x_lab}{string. The label for the x-axis. Default is \code{x_metric}.}

\item{y_lab}{string. The label for the y-axis. Default is \code{y_metric}.}

\item{width}{numeric. The width of the plot in \code{units}. Default is \code{7}.}

\item{height}{numeric. The height of the plot in \code{units}. Default is \code{4}.}

\item{units}{string. The units of \code{width} and \code{height}. Default is \code{"in"} (inches).}

\item{title}{string. The title of the plot. Default is \code{NULL}.}

\item{colors}{character vector. The colors to be used for the different models.
Default is \code{NULL}, which means that the default colors of \code{ggplot2} will be used.}

\item{show_plots}{logical. Whether to show the plots after creation in an interactive
graphics device. Default is \code{FALSE}.}
}
\value{
A PlotSpec S3 object.
}
\description{
A PlotSpec object holds all the info on how to create a model performance
plot and what to do with it. It is tailored for \code{assess_model()}. Its base object is a
list containing:
}
