% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{predict.ptk_zerosum}
\alias{predict.ptk_zerosum}
\title{Wrap \code{\link[zeroSum:predict.zeroSum]{zeroSum::predict.zeroSum()}} into a patroklos-compliant predict function}
\usage{
\method{predict}{ptk_zerosum}(object, newx, ...)
}
\arguments{
\item{object}{A ptk_patroklos S3 object.}

\item{newx}{Predictor data as a named numeric matrix. Samples correspond to
rows, an attribute \code{li_var_suffix} is expected to be present.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A numeric vector with the prediction for every sample. Unline
\code{zeroSum::predict.zeroSum()},
\itemize{
\item if \code{object$type == 2} (i.e., \code{object} was fitted with \verb{family = "binomial" by }ptk_zerosum`), the predictions are binomial probabilities,
\item if \code{object$type == 4} (i.e., \code{object} was fitted with \verb{family = "cox" by }ptk_zerosum\verb{), the predictions are the exponential of the linear predictor,  i.e., }exp(beta_0 + beta^T x)`.
\item If \code{object$binarizePredictions} is not \code{NULL}, the predictions are binarized
via \code{as.numeric(y > object$binarizePredictions)}.
}
}
\description{
\code{Model$predict()} uses this function as a predict method for
the \code{predict()} generic.
}
