% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data for model fitting and predicting}
\usage{
prepare(
  directory,
  model,
  include_from_continuous_pheno = NULL,
  include_from_discrete_pheno = NULL,
  expr_fname = "expr.csv",
  pheno_fname = "pheno.csv",
  patient_id_col = "patient_id",
  gene_id_col = "gene_id",
  pfs_leq = 2
)
}
\arguments{
\item{directory}{string. The directory where both expression and pheno csv
files lie.}

\item{model}{string. For which model to prepare. One of
\itemize{
\item \code{"cox_lasso_zerosum"} (Cox proportional hazards with LASSO reularization
and zero-sum constraint)
\item \code{"lasso_zerosum"} (ordinary linear regression with LASSO regularization)
}}

\item{include_from_continuous_pheno}{vector of strings. The names of the
\emph{continuous} variables in the pheno data file to be included in the
predictor matrix. The values will be coerced to numeric. Default is \code{NULL},
which means no continuous pheno variables will be included.}

\item{include_from_discrete_pheno}{vector of strings. The names of the
\emph{discrete} variables in the pheno data file to be included in the predictor
matrix. A discrete variable with n levels will be converted to n-1 binary
variables. Default is \code{NULL}, which means no discrete pheno variables will
be included.}

\item{expr_fname}{string. The name of the expression .csv file inside \code{directory}.
Default is \code{"expr.csv"}.}

\item{pheno_fname}{string. The name of the pheno data .csv inside \code{directory}.
Default is \code{"pheno.csv"}.}

\item{patient_id_col}{string. The name of the column in the pheno data file
that holds the patient identifiers. Default is \code{"patient_id"}.}

\item{gene_id_col}{string. The name of the column in the expression data file
that holds the gene identifiers. Default is \code{"gene_id"}.}

\item{pfs_leq}{numeric. Only necessary if model discretizes response (PFS). The value
of progression-free survival (PFS) above which samples are considered high-risk.
Default is 2.0.}
}
\value{
A list with two elements, \code{x} and \code{y}. \code{x} is the predictor matrix, \code{y} is
the response matrix with the samples as rows.
}
\description{
Read data from files, generate the predictor and
respondent matirx and control quality.
}
\details{
The pheno csv file holds the samples as rows, the variables as columns.
We need at least the columns
\itemize{
\item progression (integer, 0 for censored, 1 for uncensored),
\item pfs_yrs (integer, the time in years to progression or censoring),
\item \code{ìnclude_from_continuous_pheno},
\item \code{include_from_discrete_pheno}, and
\item \code{patient_id_col} with \emph{unique} patient identifiers.
The expr csv file holds the genes as rows (with \emph{unique} gene ids in a row called
\code{gene_id_col}), the samples as columns.
}
}
