% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data for model fitting and predicting}
\usage{
prepare(expr_mat, pheno_tbl, data_spec, model_spec)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression data, with patients as rows
and genes as columns. The row names must be unique patient identifiers.}

\item{pheno_tbl}{tibble. The pheno data, with patients as rows and variables as
columns.}

\item{data_spec}{DataSpec S3 object. Specifications on the data (\code{expr_mat},
\code{pheno_tbl}). See the the constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec S3 object. Specifications on the model to prepare for.
See the the constructor \code{ModelSpec()} for details.}
}
\value{
A list with two numeric matrices, \code{x} and \code{y}. \code{x} is the predictor matrix,
\code{y} is the response matrix, both ready to for model fitting or predicting.
}
\description{
Provided expression matrix and pheno tibble, generate the
predictor and response matrix response-type-dependently and do checks.
}
