% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_predict.R
\name{prepare_and_predict}
\alias{prepare_and_predict}
\title{Prepare data and predict with a loaded model}
\usage{
prepare_and_predict(expr_mat, pheno_tbl, data_spec, model_spec)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix with genes in rows and samples
in columns.}

\item{pheno_tbl}{tibble. The pheno data with samples in rows and variables in columns.}

\item{data_spec}{DataSpec S3 object. Specifications on the data. See the the
constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec S3 object. Specifications on the model. See the the
constructor \code{ModelSpec()} for details.}
}
\value{
A list with two elements:
\itemize{
\item \code{"predictions"}, a numeric one-column matrix holding predicted scores, and
\item "y", a numeric vector holding the true response.
}
}
\description{
Given an expression matrix and a pheno tibble, prepare the data for a
certain model and predict with this model.
}
