% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_predict.R
\name{prepare_and_predict}
\alias{prepare_and_predict}
\title{Prepare data and predict with a loaded model across splits}
\usage{
prepare_and_predict(
  expr_mat,
  pheno_tbl,
  data_spec,
  model_spec,
  lambda,
  pivot_time_cutoff,
  benchmark_col = NULL
)
}
\arguments{
\item{expr_mat}{numeric matrix. The expression matrix with genes in rows and samples
in columns.}

\item{pheno_tbl}{tibble. The pheno data with samples in rows and variables in columns.}

\item{data_spec}{DataSpec S3 object. Specifications on the data. See the the
constructor \code{DataSpec()} for details.}

\item{model_spec}{ModelSpec S3 object. Specifications on the model. See the the
constructor \code{ModelSpec()} for details.}

\item{lambda}{string or numeric. The lambda regularization parameter of the model
to predict with. Technically, we will pass it to the \code{s} parameter of the \code{predict()}
method of the object returned by the \code{fitter} attribute of the \code{ModelSpec} object.
See, e.g., \code{\link[zeroSum:predict.zeroSum]{zeroSum::predict.zeroSum()}}.}

\item{pivot_time_cutoff}{numeric. Time-to-event threshold that divides samples into a
high/low-risk (time to event below/above \code{pivot_time_cutoff}) group. See return value
below.}

\item{benchmark_col}{string or NULL. Column in the pheno tibble holding the numeric
benchmark values. If NULL, no benchmark is provided.}
}
\value{
A list holding:
\itemize{
\item \code{"predicted"}: a list of named numeric vectors, the scores output by the model for
each split (split index corresponding to list index).
\item "actual": a list of named numeric vectors, for each split the actual values of
whether time to event was above or below \code{model_spec$time_cutoffs}, encoded as 1
("high risk") and 0 ("low risk"), respectively.
\item "benchmark": A list of named numeric vectors, for each split the values of the
benchmark classifier. If \code{benchmark} is NULL, it is an empty list.
For every split, the names of all three vectors match.
}
}
\description{
Given an expression matrix and a pheno tibble, prepare the data for a
certain model and predict with this model, provide the actual and optionally the
benchmark values. Do this for all splits into training and test cohort. As with
\verb{[prepare_and_fit()]}, we do not support multiple time cutoffs at this step; this
is \verb{[assess_2d_center()]}'s job.
}
