% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{prepend_to_directory}
\alias{prepend_to_directory}
\title{To every \code{Model} in a list of \code{Model}s, prepend a fixed directory to the
\code{directory} attribute}
\usage{
prepend_to_directory(model_list, dir)
}
\arguments{
\item{model_list}{list of \code{Model}s.}

\item{dir}{string. The directory to prepend to the \code{directory} attribute of all
\code{Model}s in \code{model_list}.}
}
\value{
A list of cloned \code{Model}s, with \code{dir} prepended to the \code{directory} attribute.
}
\description{
Often one specifies the models in general, for all data sets.
If you fit the models to a specific data set (say \code{"mock"}), you might want
to prepend a fixed directory like \code{"results/mock"} to the \code{directory}
attribute of all \code{Model}s in the list.
}
