% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{ptk_ranger}
\alias{ptk_ranger}
\title{Wrap \code{\link[ranger:ranger]{ranger::ranger()}} into a patroklos-compliant fitter}
\usage{
ptk_ranger(x, y, rel_mtry, mtry = NULL, skip_on_invalid_input = FALSE, ...)
}
\arguments{
\item{x}{Named numeric matrix. Predictor matrix without \code{NA}s. Samples correspond to rows.
Discrete features are encoded as binary dummy variables.}

\item{y}{Named list with the response in thee formats:
\itemize{
\item \code{"bin"}, a named numeric one-column matrix, binary response,
\item \code{"cox"}, a named numeric two-column matrix, with columns \code{"time_to_event"} and \code{"event"}
(0 = censoring, 1 = event), the response in the Cox format,
\item \code{"true"}, a named numeric one-column matrix, true binary response.
}

The rownames of \code{y[["bin"]]} and \code{y[["true"]]} are a subset of the rownames of \code{x} and, in
general, do not coincide. Use \code{intersect_by_names()} to get equal rownames.}

\item{rel_mtry}{logical. If \code{TRUE}, interpret \code{mtry} as relative to
\code{sqrt(ncol(x))} (the recommended value), rounded to the next integer.
Otherwise, take \code{mtry} directly.}

\item{skip_on_invalid_input}{Logical. If \code{TRUE} and invalid input is detected,
return \code{NA} instead of an error. This is useful when calling this function
from inside \code{\link[=long_nestor]{long_nestor()}}.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A \code{ptk_ranger} S3 object, a \code{ranger} S3 object with the (OOB)
\code{predictions} attribute renamed to \code{val_predict}.
}
\description{
This function is a patroklos-compliant fitter with validated
predictions, it has a return value with a \code{val_predict} attribute.
}
