% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{ptk_ranger}
\alias{ptk_ranger}
\title{Wrap \code{\link[ranger:ranger]{ranger::ranger()}} into a patroklos-compliant fitter}
\usage{
ptk_ranger(x, y, ...)
}
\arguments{
\item{x}{Predictor data as a named numeric matrix. Samples correspond to rows,
an attribute \code{li_var_suffix} is expected to be present.}

\item{y}{Response data as numeric vector. The length of \code{y} must match the
number of rows in \code{x}.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A \code{ptk_ranger} S3 object, a \code{ranger} S3 object with the \code{predictions}
attribute renamed to \code{oob_predict}.
}
\description{
This function is a patroklos-compliant fitter with validated
predictions, it has a return value with a \code{oob_predict} attribute.
}
