% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{ptk_ranger}
\alias{ptk_ranger}
\title{Wrap \code{\link[ranger:ranger]{ranger::ranger()}} into a patroklos-compliant fitter}
\usage{
ptk_ranger(x, y, mtry = NULL, skip_on_invalid_input = FALSE, ...)
}
\arguments{
\item{x}{Predictor data (independent variables), alternative interface to data with formula or dependent.variable.name.}

\item{y}{Response vector (dependent variable), alternative interface to data with formula or dependent.variable.name. For survival use a \code{Surv()} object or a matrix with time and status.}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables. Alternatively, a single argument function returning an integer, given the number of independent variables.}

\item{skip_on_invalid_input}{Logical. If \code{TRUE} and invalid input is detected,
return \code{NA} instead of an error. This is useful when calling this function
from inside \code{\link[=nested_pseudo_cv]{nested_pseudo_cv()}}.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A \code{ptk_ranger} S3 object, a \code{ranger} S3 object with the (OOB)
\code{predictions} attribute renamed to \code{val_predict}.
}
\description{
This function is a patroklos-compliant fitter with validated
predictions, it has a return value with a \code{val_predict} attribute.
}
