% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{ptk_zerosum}
\alias{ptk_zerosum}
\title{Wrap \code{\link[zeroSum:zeroSum]{zeroSum::zeroSum()}} into a patroklos-compliant fit function}
\usage{
ptk_zerosum(
  x,
  y,
  exclude_pheno_from_lasso = TRUE,
  binarize_predictions = NULL,
  ...,
  zeroSum.weights = NULL,
  penalty.factor = NULL
)
}
\arguments{
\item{x}{data as a numeric matrix object (rows=samples). The zero-sum
regression requires data on the log scale, i.e. x should be
log-transformed data.}

\item{y}{response vector to be predicted by x (length(y)==nrow(x))}

\item{exclude_pheno_from_lasso}{Logical. If \code{TRUE}, set LASSO penalty weights
corresponding to features from the pheno data to zero.}

\item{binarize_predictions}{numeric or NULL. If not NULL, the predict method
for the returned \code{ptk_zerosum} object will binarize the predictions using the
\code{binarize_predictions} as a threshold.}

\item{...}{can be used for adjusting internal parameters}

\item{zeroSum.weights}{weights vector for the zero-sum constraint of
length ncol(x). By setting a weight to 0 the corresponding feature
will be excluded from the zero-sum constraint.
(must be greater than or equal to zero, default 1)}

\item{penalty.factor}{weights vector for the elatic net regularization of
length ncol(x). By setting a weight to 0 the corresponding feature
will not be regularized and thus will be part of the resulting model.
(must be greater than or equal to zero, default 1).}
}
\value{
A \code{ptk_zerosum} S3 object.
}
\description{
You can now use it as the \code{fitter} attribute of a \code{Model} object.
This fitter is suited for early integration.
}
