% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{ptk_zerosum}
\alias{ptk_zerosum}
\title{Wrap \code{\link[zeroSum:zeroSum]{zeroSum::zeroSum()}} into a patroklos-compliant fit function}
\usage{
ptk_zerosum(
  x,
  y,
  exclude_pheno_from_lasso = TRUE,
  binarize_predictions = NULL,
  ...,
  nFold = 10,
  zeroSum.weights = NULL,
  penalty.factor = NULL,
  family = "binomial"
)
}
\arguments{
\item{x}{data as a numeric matrix object (rows=samples). The zero-sum
regression requires data on the log scale, i.e. x should be
log-transformed data.}

\item{y}{response vector to be predicted by x (length(y)==nrow(x))}

\item{exclude_pheno_from_lasso}{Logical. If \code{TRUE}, set LASSO penalty weights
corresponding to features from the pheno data to zero.}

\item{binarize_predictions}{numeric or NULL. If not NULL, the predict method
for the returned \code{ptk_zerosum} object will binarize the predictions using the
\code{binarize_predictions} as a threshold.}

\item{...}{can be used for adjusting internal parameters}

\item{nFold}{the number of folds used by the cross validation (Default 10)}

\item{zeroSum.weights}{weights vector for the zero-sum constraint of
length ncol(x). By setting a weight to 0 the corresponding feature
will be excluded from the zero-sum constraint.
(must be greater than or equal to zero, default 1)}

\item{penalty.factor}{weights vector for the elatic net regularization of
length ncol(x). By setting a weight to 0 the corresponding feature
will not be regularized and thus will be part of the resulting model.
(must be greater than or equal to zero, default 1).}

\item{family}{choose the regression type:
\describe{
    \item{gaussian:}{numeric response}
    \item{binomial:}{}
    \item{multinomial:}{}
    \item{cox:}{y should be matrix with two columns, the first must
                contain the event time and the second indicating the
                 type: 1 = event has occured, 0 = right censoring}
}}
}
\value{
A \code{ptk_zerosum} S3 object.
}
\description{
This function is a patroklos-compliant fitter with integrated
CV and, if \code{length(lambda) == 1}, also a patroklos-compliant fitter with
validated predictions.
}
\details{
A \emph{patroklos-compliant fitter with integrated CV} is a patroklos-
compliant fitter (cf. details of \code{\link[=ptk_ranger]{ptk_ranger()}}) whose return value has an
attribute \code{cv_predict_list}, a list of numeric vectors holding cross-validated
predictions for every value of the model hyperparameter lambda.
}
