% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitters.R
\name{ptk_zeroSum}
\alias{ptk_zeroSum}
\title{Wrap \code{\link[zeroSum:zeroSum]{zeroSum::zeroSum()}} into a patroklos-compliant fit function}
\usage{
ptk_zeroSum(
  x,
  y,
  exclude_pheno_from_lasso = TRUE,
  binarize_predictions = NULL,
  ...
)
}
\arguments{
\item{x}{Predictor data as a named numeric matrix. Samples correspond to rows,
an attribute \code{li_var_suffix} is expected to be present.}

\item{y}{Response data as numeric vector. The length of \code{y} must match the
number of rows in \code{x}.}

\item{exclude_pheno_from_lasso}{Logical. If \code{TRUE}, set LASSO penalty weights
corresponding to features from the pheno data to zero.}

\item{binarize_predictions}{numeric or NULL. If not NULL, the predict method
for the returned \code{ptk_zeroSum} object will binarize the predictions using the
\code{binarize_predictions} as a threshold.}

\item{...}{Further arguments passed to the wrapped function.}
}
\value{
A \code{ptk_zeroSum} S3 object.
}
\description{
You can now use it as the \code{fitter} attribute of a \code{Model} object.
This fitter is suited for early integration.
}
