% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_preprocess}
\alias{qc_preprocess}
\title{Quality control at the end of preprocessing step}
\usage{
qc_preprocess(data, expr_tbl, pheno_tbl)
}
\arguments{
\item{data}{A \code{Data} object.}

\item{expr_tbl, pheno_tbl}{A tibble with the expression and pheno data.}
}
\description{
Check if the expression and pheno tibble have the format
specified in the \code{Data} and match to one another.
}
\details{
Intended to be used at the end of preprocessing.
}
