% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read expression and pheno data from csv files into consistent format}
\usage{
read(directory, expr_fname, pheno_fname, patient_id_col, gene_id_col)
}
\arguments{
\item{directory}{string. The directory where both expression and pheno csv
files lie.}

\item{expr_fname}{string. The name of the expression .csv file inside \code{directory}.
See details for the expected format.}

\item{pheno_fname}{string. The name of the pheno data .csv inside \code{directory}.
See details for the expected format.}

\item{patient_id_col}{string. The name of the column in the pheno data file
that holds the patient identifiers.}

\item{gene_id_col}{string. The name of the column in the expression data file
that holds the gene identifiers.}
}
\value{
A list with a numeric matrix, named \code{expr}, and a tibble named \code{pheno}.
\code{expr} holds the expression data, with patient ids as row names and gene ids as
column names. I.e., we transpose the expression data. \code{pheno} holds the pheno data,
with the patient ids in the first column \code{patient_id_col}.
}
\description{
Read expression data into a matrix and pheno data into a tibble.
Both will hold patients as rows.
}
\details{
The pheno csv file holds the samples as rows (with the unique sample names
in the column \code{patient_id}), the variables as columns. We need at least the columns
\itemize{
\item progression (integer, 0 for censored, 1 for uncensored),
\item pfs_yrs (integer, the time in years to progression or censoring),
\item \code{ìnclude_from_continuous_pheno},
\item \code{include_from_discrete_pheno} and
\item \code{patient_id_col}.
The expr csv file holds the genes as rows (with the gene ids in a row called
\code{gene_id_col}), the samples as columns. The sample names in pheno and expression files
must be identical and in the same order.
}
}
