% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_camp.R
\name{training_camp}
\alias{training_camp}
\title{Set up a training camp to fit models}
\usage{
training_camp(data_spec, model_spec_list)
}
\arguments{
\item{data_spec}{DataSpec object. The data set to fit on.}

\item{model_spec_list}{list of ModelSpec objects. The models to fit.}
}
\description{
Given one data set and a list of models, fit all models for all
splits and time cutoffs. Store the models.
}
