% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_vs_test.R
\name{val_vs_test}
\alias{val_vs_test}
\title{Compare validation and test error}
\usage{
val_vs_test(
  model_list,
  data,
  error_fun,
  spotlight_regex,
  file,
  spotlight_name = NULL,
  plot_theme = NULL,
  quiet = FALSE
)
}
\arguments{
\item{model_list}{list of Model objects.}

\item{data}{Data object.}

\item{error_fun}{function. Error function to compare validation and test error.}

\item{spotlight_regex}{string. Regular expression to identify models to
highlight.}

\item{file}{string. File name to save the plot to.}

\item{spotlight_name}{string. Name to give to the models identified by
\code{spotlight_regex} in legend.}

\item{plot_theme}{ggplot2 theme. Theme to apply to the plot.}

\item{quiet}{logical. Whether to suppress messages.}
}
\value{
ggplot2 object.
}
\description{
Plot rank of validation error versus test error - validation
error for a list of models.
}
