% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{write_data_info}
\alias{write_data_info}
\title{Write data info to a JSON file}
\usage{
write_data_info(filename, data, expr_tbl)
}
\arguments{
\item{filename}{string. The path to the JSON file.}

\item{data}{\code{Data} object. Analyzed data.}

\item{expr_tbl}{tibble. The expression data. Rows are genes, columns are samples. The first
column, named \code{data$gene_id_col}, holds the gene identifiers.}
}
\value{
The JSON string that was written to the file.
}
\description{
Automatically generate info about the data and write it to a JSON file.
If the file already exists, write the automatically inferable info into the file. If the
file is not present yet, create a skeleton, write the automatically inferable info into it. You
can then fill the empty fields in by writing the json by hand. Manually already filled out
info will not be overwritten.
}
